### 2020-08-12 #395 @agila5
Submission: osmextract
Submitting Author: Andrea Gilardi (@agila5)
Repository:  https://github.com/ITSLeeds/osmextract
Version submitted:  0.1.0
Version accepted: TBD

---

```
Package: osmextract
Title: Download and Read OpenStreetMap Data Extracts
Version: 0.1.0
    person("Andrea", "Gilardi", email = "a.gilardi5@campus.unimib.it", role=c("aut", "cre"),
    person("Robin", "Lovelace", role = c("aut"),
    person("Barry", "Rowlingson", role=c("ctb"),
    )
Description: This package is used to download, convert and read Open Street
    Map data extracts downloaded from several providers.
    sf (>= 0.8.1),
    tools
    covr
```

`osmextract` is for accessing and downloading OpenStreetMap data from a variety (current 2) of providers. OpenStreetMap has numerous scientific applications, including in transport, ecological analysis of human impacts on the environment, spatial statistics and land use change and the package provides a way of extracting the unstructure compressed `.pbf` files into modern geographic data structures (`sf` data frames).

The target audience is researchers, practitioners and citizens who want to use OpenStreetMap data for scientific research, especially people who require large datasets (from city to continental scale) for their work.

[`osmdata`](https://github.com/ropensci/osmdata) provides an R interface to the [Overpass API](https://wiki.openstreetmap.org/wiki/Overpass_API),  which is ideal for downloading small OSM datasets. However, the API is rate limited, making it hard to download large datasets.

We believe the package complies with ethical guidelines in the [Internet Research: Ethical Guidelines 3.0 document](https://aoir.org/reports/ethics3.pdf).
It makes it easier for researchers to access and make use of data that is already in the public domain, under the conditions of the adhering to the conditions of the [OdBL](https://opendatacommons.org/licenses/odbl/).
