### 2020-08-13 8cb3802
update osmconf_ini in vignette skip ci
