+++ vignettes/osmextract.Rmd
@@ -393,7 +393,7 @@ The following example shows how to create an ad-hoc `osmconf.ini` file, which is
 custom_osmconf_ini <- readLines(system.file("osmconf.ini", package = "osmextract"))
 ```
 
-Then we modify the code at lines 18 and 21, setting the parameters for reporting all nodes and ways even without any significant tag:
+Then we modify the code at lines 18 and 21, setting the parameters for reporting all nodes and ways even without any significant tag,
 
 ```{r}
 custom_osmconf_ini[[18]] = "report_all_nodes=yes"
@@ -407,7 +407,7 @@ custom_osmconf_ini[[45]] = "osm_id=no"
 custom_osmconf_ini[[53]] = "attributes=highway,lanes"
 ```
 
-Another important parameter that can be modified creating a custom `osmconf.ini` file is `closed_ways_area_polygons` (see line 7 of the default CONFIG file). 
+Another important parameter that can be customized creating an ad-hoc `osmconf.ini` file is `closed_ways_area_polygons` (see lines 5-7 of the default CONFIG file). 
 We can now save the `custom_osmconf_ini` file: 
 
 ```{r}
@@ -422,20 +422,20 @@ oe_get("ITS Leeds", provider = "test", quiet = FALSE)
 oe_get("ITS Leeds", provider = "test", osmconf_ini = temp_ini, quiet = FALSE)
 ```
 
-We can see that there are 3 extra features in the `sf` data frame read-in using the customized CONFIG file (since we set `"report_all_nodes=yes"` and `"report_all_ways=yes"`) and just 4 field: `highway`, `lanes`, (see the code a few chunks above), `z_order` (see [here](https://github.com/OSGeo/gdal/blob/9f31018839b32aeeafad7663a8de662153a956c3/gdal/data/osmconf.ini#L65-L71)), and `other_tags`.
+We can see that there are 3 extra features in the `sf` object that was read-in using the customized CONFIG file (since we set `"report_all_nodes=yes"` and `"report_all_ways=yes"`) and just 4 field: `highway`, `lanes`, (see the code a few chunks above), `z_order` (see [here](https://github.com/OSGeo/gdal/blob/9f31018839b32aeeafad7663a8de662153a956c3/gdal/data/osmconf.ini#L65-L71)), and `other_tags`.
 
 ### `vectortranslate_options`
 
 The parameter `vectortranslate_options` is used to modify the options that are passed to [ogr2ogr](https://gdal.org/programs/ogr2ogr.html#ogr2ogr). 
-This is extremely important because if we tune the `vectortranslate_options` parameter, then we can analyze big `.pbf` files without reading them in memory. 
+This is extremely important because if we tune the  options of `vectortranslate_options` parameter, then we can analyze big `.pbf` files without fully reading them in memory. 
 
-The first example is reported in the following chunk that shows how to use the argument `-t_srs` to modify the CRS of the output `.gpkg` object while performing vectortranslate operations: 
+The first example is reported in the following chunk of code and it shows how to use the argument `-t_srs` to modify the CRS of the output `.gpkg` object (from `EPSG:4326` to [`EPSG:27700`](https://epsg.io/27700)) while performing vectortranslate operations: 
 
 ```{r}
 my_osmconf_ini = system.file("osmconf.ini", package = "osmextract")
 my_vectortranslate = c(
   "-f", "GPKG", #output file format
-  "-overwrite", # overwrite an existing file
+  "-overwrite", # overwrite an existing layer
   "-oo", paste0("CONFIG_FILE=", my_osmconf_ini), # open options
[…]
