### 2020-08-14 a0f4064
First draft of introductory vignette.
Fix #96
