+++ vignettes/osmextract.Rmd
@@ -23,7 +23,7 @@ Loading the package generates important messages about the license associated wi
 library(osmextract)
 ```
 
-The first thing to say is: do not ignore this message!
+The first thing to say is: **do not ignore this message**!
 There are important legal considerations that you should be aware of before using OSM data, especially if you are working in a for-profit capacity.
 
 # Legal considerations
@@ -42,11 +42,11 @@ See the [License/Use Cases page on the OSM wiki](https://wiki.openstreetmap.org/
 The packages is composed of the following main functions: 
 
 1. `oe_providers()`: Show which providers of OSM extracts are available;
-1. `oe_match()`: Match the input zone with one of the files stored by the OSM providers;
+1. `oe_match()`: Match one input place with one of the files stored by the OSM providers;
 1. `oe_download()`: Download the chosen file;
-1. `oe_vectortranslate()`: Convert `.pbf` files into `.gpkg` files;
+1. `oe_vectortranslate()`: Convert between `.pbf` and `.gpkg` formats;
 1. `oe_read()`: Read `.pbf` and `.gpkg` files;
-1. `oe_get()`: Match, download, translate and import data in one step.
+1. `oe_get()`: Match, download, translate and import data, all in one step.
 
 For many users who just want to get OSM data quickly `oe_get()` may be sufficient, as covered in the README.
 We will demonstrate each function in turn.
@@ -59,8 +59,9 @@ We will demonstrate each function in turn.
 oe_providers()
 ```
 
-Each element in `database_name` is a data object that is packaged with `osmextract`. 
+Each element in the column `database_name` is a data object that is packaged with `osmextract`. 
 You can read a detailed description of each provider data running, for example, `?geofabrik_zones` or `?bbbike_zones`. 
+
 Perhaps the best known bulk OSM data provider is Geofabrik, represented as data frame in the packaged object `geofabrik_zones`.
 
 ```{r}
@@ -95,11 +96,12 @@ You can also download manually selected regions of interest from the BBBike webs
 
 ## `oe_match()`: finding OSM extracts
 
-`oe_match()` takes one character string and returns one matching zone based on the minimum Approximate String Distance (`?adist`) between the input `place` and the `name` column in the provider data, such as `geofabrik_zones` (the default).
-By default it returns a named list with the URL and file size (in bytes) of the corresponding `pbf` file hosted by the provider, as shown below.
+`oe_match()` takes one character string representing a `place` and returns one matching zone based on the minimum Approximate String Distance (`?adist`) between the input `place` and the names of the zones of the provider. 
+The matching is performed using the `name` field of the provider data and the default provider is Geofabrik.
+The function returns a named list with the URL and file size (in bytes) of the corresponding `pbf` file hosted by the provider, as shown below.
 
 ```{r}
-oe_match("Italy")
+oe_match("Italy", provider = "geofabrik")
 oe_match("Leeds", provider = "bbbike")
 ```
 
@@ -164,8 +166,8 @@ oe_match("Belin")
 The input `place` can be also specified using an `sfc_POINT` object with arbitrary CRS as documented in the following example. 
 The function will return a zone intersecting the `sfc_POINT` object (or an error, if the input point does not intersect any area). 
 If the input `place` intersects multiple geographically nested zones, the function returns the area with the highest `level`. 
-Check the help page of `?geofabrik_zones` to understand the meaning of the `level` field.  
[…]
