### 2020-08-14 7382c36
Set quiet = FALSE for large files in download + typo
