+++ man/oe_update.Rd
@@ -67,7 +67,7 @@ Check \code{\link[=file.info]{file.info()}}.
 \examples{
 \dontrun{
 # Set up a fake directory with .pbf and .gpkg files
-fake_dir <- tempdir()
+fake_dir = tempdir()
 # Fill the directory
 oe_get("andorra", download_directory = fake_dir, download_only = TRUE)
 # Check the directory
