### 2020-08-14 fa46594
typo and some things from goodpractice
