+++ DESCRIPTION
@@ -23,8 +23,7 @@ Depends: R (>= 3.5.0)
 Imports: 
     sf (>= 0.8.1),
     utils,
-    tools,
-    s2
+    tools
 Suggests: 
     testthat,
     knitr,
