+++ man/oe_file.Rd
@@ -1,0 +1,44 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/file.R
+\name{oe_file}
+\alias{oe_file}
+\title{Get the location of files downloaded by osmextractr}
+\usage{
+oe_file(
+  place,
+  download = FALSE,
+  download_directory = oe_download_directory(),
+  ...
+)
+}
+\arguments{
+\item{place}{Description of the geographical area that should be matched with
+a \code{.osm.pbf} file through the chosen \code{provider}. Can be either a length-1
+character vector, a length-1 \code{sfc_POINT} object or a numeric vector of
+coordinates with length 2. In the latter case it is assumed that the EPSG
+code is 4326, while you can use any EPSG code with an \code{sfc_POINT} object.
+See Details and examples in \code{\link[=oe_match]{oe_match()}}.}
+
+\item{download}{Attempt to download the file if it cannot be found?
+\code{FALSE} by default.}
+
+\item{download_directory}{Where to download the file containing the OSM data?
+By default this is equal to \code{\link[=oe_download_directory]{oe_download_directory()}}, which is equal to
+\code{tempdir()} and it changes each time you restart R. You can set a
+persistent \code{download_directory} by adding the following to your \code{.Renviron}
+file (e.g. with \code{\link[usethis:edit]{usethis::edit_r_environ()}}):
+\verb{OSMEXT_DOWNLOAD_DIRECTORY=/path/to/osm/data}.}
+
+\item{...}{Arguments that will be passed to \code{\link[sf:st_read]{sf::st_read()}}, like \code{query}
+or \code{stringsAsFactors}.  Check the introductory vignette to understand how
+to create your own (SQL-like) queries.}
+}
+\description{
+This function takes a place name and
+}
+\examples{
+res = oe_get("ITS", provider = "test", max_string_dist = 9)
+oe_file("its")
+oe_file("Isle of Wight")
+oe_file("Non-existant-place")
+}
