### 2020-09-06 4665cd8
Update ci for sf + new OSGeo libs
