+++ .github/workflows/R-CMD-check.yaml
@@ -86,6 +86,7 @@ jobs:
         run: |
           remotes::install_deps(dependencies = TRUE)
           remotes::install_cran("rcmdcheck")
+          install.packages("sf") # reinstall sf matching new OSGeo packages
         shell: Rscript {0}
 
       - name: Session info
