### 2020-09-06 b9ad925
Reinstall sf from CRAN
