+++ man/oe_find.Rd
@@ -1,0 +1,48 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/file.R
+\name{oe_find}
+\alias{oe_find}
+\title{Get the location of files downloaded by osmextractr}
+\usage{
+oe_find(
+  place,
+  download = FALSE,
+  download_directory = oe_download_directory(),
+  ...
+)
+}
+\arguments{
+\item{place}{Description of the geographical area that should be matched with
+an existing \code{.pbf} and \code{.gpkg} file. Must be a character vector of length
+one.}
+
+\item{download}{Attempt to download the file if it cannot be found?
+\code{FALSE} by default.}
+
+\item{download_directory}{Directory where the files downloaded by osmextract
+are stored. By default it is equal to \code{oe_download_directory()}.}
+
+\item{...}{Extra arguments that are passed to \code{oe_get()} in case \code{download}
+is \code{TRUE} and the input \code{place} is matched with no existing file. Please
+note that you cannot modify the argument \code{download_only}.}
+}
+\description{
+This function takes a \code{place} name and it returns the path of \code{.pbf} and
+\code{.gpkg} files associated with it
+}
+\details{
+The matching operation between the existing files (downloaded by
+osmextract and stored in \code{oe_download_directory()}) and the input \code{place}
+is performed using a regular expression with a pattern equal to the input
+\code{place}. If the function finds no file associated with the input \code{place}
+and \code{download} parameter is equal to \code{TRUE}, then it tries to download and
+translate a new file from the chosen provider
+}
+\examples{
+res = oe_get("ITS Leeds", provider = "test")
+oe_find("ITS")
+oe_find("Isle of Wight")
+\dontrun{
+oe_find("Isle of Wight", download = TRUE)}
+oe_find("Non-existant-place")
+}
