+++ NAMESPACE
@@ -6,7 +6,7 @@ S3method(oe_match,numeric)
 S3method(oe_match,sfc_POINT)
 export(oe_download)
 export(oe_download_directory)
-export(oe_file)
+export(oe_find)
 export(oe_get)
 export(oe_get_keys)
 export(oe_match)
