### 2020-09-08 a819134
new approach for oe_find
