+++ man/oe_find.Rd
@@ -1,48 +1,62 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/file.R
+% Please edit documentation in R/find.R
 \name{oe_find}
 \alias{oe_find}
 \title{Get the location of files downloaded by osmextractr}
 \usage{
 oe_find(
   place,
-  download = FALSE,
+  provider = "geofabrik",
   download_directory = oe_download_directory(),
+  download_if_missing = FALSE,
   ...
 )
 }
 \arguments{
 \item{place}{Description of the geographical area that should be matched with
-an existing \code{.pbf} and \code{.gpkg} file. Must be a character vector of length
-one.}
+a \code{.osm.pbf} file through the chosen \code{provider}. Can be either a length-1
+character vector, a length-1 \code{sfc_POINT} object or a numeric vector of
+coordinates with length 2. In the latter case it is assumed that the EPSG
+code is 4326, while you can use any EPSG code with an \code{sfc_POINT} object.
+See Details and examples in \code{\link[=oe_match]{oe_match()}}.}
 
-\item{download}{Attempt to download the file if it cannot be found?
-\code{FALSE} by default.}
+\item{provider}{Which provider should be used to download the data? Available
+providers can be found with the following command: \code{\link[=oe_providers]{oe_providers()}}.}
 
 \item{download_directory}{Directory where the files downloaded by osmextract
 are stored. By default it is equal to \code{oe_download_directory()}.}
 
-\item{...}{Extra arguments that are passed to \code{oe_get()} in case \code{download}
-is \code{TRUE} and the input \code{place} is matched with no existing file. Please
-note that you cannot modify the argument \code{download_only}.}
+\item{download_if_missing}{Attempt to download the file if it cannot be found? \code{FALSE}
+by default.}
+
+\item{...}{Extra arguments that are passed to \code{oe_match} and \code{oe_get()}.
+Please note that you cannot modify the argument \code{download_only}.}
 }
 \description{
 This function takes a \code{place} name and it returns the path of \code{.pbf} and
-\code{.gpkg} files associated with it
+\code{.gpkg} files associated with it.
 }
 \details{
-The matching operation between the existing files (downloaded by
-osmextract and stored in \code{oe_download_directory()}) and the input \code{place}
-is performed using a regular expression with a pattern equal to the input
-\code{place}. If the function finds no file associated with the input \code{place}
-and \code{download} parameter is equal to \code{TRUE}, then it tries to download and
-translate a new file from the chosen provider
+The matching between the existing files (saved in a directory
+specified by \code{download_directory} parameter) and the input \code{place} is
+performed using \code{list.files}, setting a pattern equal to the basename of
+the URL associated to the input \code{place}. For example, if you specify
+\code{place = "Isle of Wight"}, then the input \code{place} is matched with a URL of
+a \code{.osm.pbf} file (via \code{oe_match}) and the matching is performed setting a
+pattern equal to the basename of that URL.
+
+If there is no file in \code{download_directory} that can be matched with the
+basename and \code{download_if_missing} parameter is equal to \code{TRUE}, then the
+function tries to download and translate a new file from the chosen
+provider (\code{geofabrik} is the default provider). If \code{download_if_missing}
+parameter is equal to \code{FALSE} (default value), then the function stops with
+an error.
 }
 \examples{
 res = oe_get("ITS Leeds", provider = "test")
-oe_find("ITS")
-oe_find("Isle of Wight")
+oe_find("ITS Leeds", provider = "test")
 \dontrun{
-oe_find("Isle of Wight", download = TRUE)}
-oe_find("Non-existant-place")
[…]
