### 2020-09-08 7ca59ac
fix tests and set quiet = TRUE
