+++ tests/testthat/test-update.R
@@ -7,7 +7,7 @@ test_that("oe_update(): simplest example works", {
     download_directory = fake_dir,
     download_only = TRUE
   )
-  expect_error(oe_update(fake_dir), NA)
+  expect_error(oe_update(fake_dir, quiet = TRUE), NA)
 })
 
 
