+++ pkgdown.yml
@@ -1,6 +1,6 @@
 pandoc: 2.7.3
-pkgdown: 1.5.1.9000
-pkgdown_sha: eae56f08694abebf93cdfc0dd8e9ede06d8c815f
+pkgdown: 1.6.0.9000
+pkgdown_sha: f3d5e4141b7893df4f0a122983f89a440d187625
 articles: []
-last_built: 2020-08-18T10:05Z
+last_built: 2020-09-08T16:18Z
 
