### 2020-09-16 5dc0e57
Shorten and tidy oe_get examples
