### 2020-09-16 0a31cd5
Add example showing query argument
