### 2020-09-16 41d1308
Add link to osmextract vignette
