+++ R/read.R
@@ -52,7 +52,7 @@ oe_read = function(
   file_path,
   layer = "lines",
   ...,
-  provider = NULL,
+  provider = "geofabrik",
   download_directory = oe_download_directory(),
   file_size = NULL,
   force_download = FALSE,
