+++ man/oe_find.Rd
@@ -21,7 +21,9 @@ code is 4326, while you can use any EPSG code with an \code{sfc_POINT} object.
 See Details and examples in \code{\link[=oe_match]{oe_match()}}.}
 
 \item{provider}{Which provider should be used to download the data? Available
-providers can be found with the following command: \code{\link[=oe_providers]{oe_providers()}}.}
+providers can be found with the following command: \code{\link[=oe_providers]{oe_providers()}}. If
+\code{place} is equal to \verb{ITS Leeds}, then \code{provider} is set equal to \code{test}.
+This is just for simple examples and internal testings.}
 
 \item{download_directory}{Directory where the files downloaded by osmextract
 are stored. By default it is equal to \code{oe_download_directory()}.}
