+++ pkgdown.yml
@@ -1,6 +1,8 @@
 pandoc: 2.7.3
-pkgdown: 1.6.0.9000
-pkgdown_sha: f3d5e4141b7893df4f0a122983f89a440d187625
-articles: []
-last_built: 2020-09-08T16:18Z
+pkgdown: 1.6.1.9000
+pkgdown_sha: cc9c7b1be8573f8af3453f6e5d0422d86c916701
+articles:
+  osmextract: osmextract.html
+  providers: providers.html
+last_built: 2020-09-16T21:06Z
 
