### 2020-09-22 dbf1316
Add openstreetmap_fr provider. Fix #116
