+++ R/globals.R
@@ -2,6 +2,7 @@ utils::globalVariables(
   c(
     "geofabrik_zones",
     "test_zones",
-    "bbbike_zones"
+    "bbbike_zones",
+    "openstreetmap_fr_zones"
     )
   )
