### 2020-09-22 201a70f
typos in bbbike_zones and geofabrik_zones
