+++ pkgdown.yml
@@ -1,8 +1,8 @@
 pandoc: 2.7.3
 pkgdown: 1.6.1.9000
-pkgdown_sha: cc9c7b1be8573f8af3453f6e5d0422d86c916701
+pkgdown_sha: e1e46f2e13c7efdb17bc4eae40bc3779c2b0b517
 articles:
   osmextract: osmextract.html
   providers: providers.html
-last_built: 2020-09-16T21:06Z
+last_built: 2020-09-22T16:39Z
 
