+++ R/get.R
@@ -17,9 +17,10 @@
 #' @param layer Which `layer` should be read in? Typically `points`, `lines`
 #' (the default), `multilinestrings`, `multipolygons` or `other_relations`.
 #' @param provider Which provider should be used to download the data? Available
-#'   providers can be found with the following command: [oe_providers()]. If
-#'   `place` is equal to `ITS Leeds`, then `provider` is set equal to `test`.
-#'   This is just for simple examples and internal testings.
+#'   providers can be found with the following command: [oe_providers()]. For
+#'   `oe_get()` and `oe_match()`, if `place` is equal to `ITS Leeds`, then
+#'   `provider` is set equal to `test`. This is just for simple examples and
+#'   internal testings.
 #' @param match_by Which column of the provider's database should be used for
 #'   matching the input `place` with a `.osm.pbf` file? The default is "name".
 #'   Check details and examples in [oe_match()] to understand how this parameter
