+++ man/oe_read.Rd
@@ -33,9 +33,10 @@ or \code{stringsAsFactors}.  Check the introductory vignette to understand how
 to create your own (SQL-like) queries.}
 
 \item{provider}{Which provider should be used to download the data? Available
-providers can be found with the following command: \code{\link[=oe_providers]{oe_providers()}}. If
-\code{place} is equal to \verb{ITS Leeds}, then \code{provider} is set equal to \code{test}.
-This is just for simple examples and internal testings.}
+providers can be found with the following command: \code{\link[=oe_providers]{oe_providers()}}. For
+\code{oe_get()} and \code{oe_match()}, if \code{place} is equal to \verb{ITS Leeds}, then
+\code{provider} is set equal to \code{test}. This is just for simple examples and
+internal testings.}
 
 \item{download_directory}{Where to download the file containing the OSM data?
 By default this is equal to \code{\link[=oe_download_directory]{oe_download_directory()}}, which is equal to
