### 2020-09-23 b190e01
Add note for progress bar on vectortranslate
Fix #124
