+++ pkgdown.yml
@@ -4,5 +4,5 @@ pkgdown_sha: e1e46f2e13c7efdb17bc4eae40bc3779c2b0b517
 articles:
   osmextract: osmextract.html
   providers: providers.html
-last_built: 2020-09-22T16:39Z
+last_built: 2020-09-23T05:52Z
 
