### 2020-09-24 b76ea5e
First draft rewriting first part of intro vignette
