+++ R/providers.R
@@ -70,7 +70,7 @@ oe_providers = function(quiet = FALSE) {
   if (isFALSE(quiet)) {
     message(
       "Check the corresponding help pages to read more details about the ",
-      "fields in each database"
+      "fields in each database (e.g. ?geofabrik_zones)"
     )
   }
 
