+++ vignettes/osmextract.Rmd
@@ -28,18 +28,18 @@ There are important legal considerations that you should be aware of before usin
 
 # Legal considerations
 
-Anyone using OSM data is bound by law to adhere to the [ODbL](https://opendatacommons.org/licenses/odbl/summary/) which means that you must:
+Anyone using OSM data is bound by law to adhere to the [ODbL](https://opendatacommons.org/licenses/odbl/summary/), which means that you must:
 
-- Attribute: You must attribute any public use of the database, or works produced from the database, in the manner specified in the ODbL. For any use or redistribution of the database, or works produced from it, you must make clear to others the license of the database and keep intact any notices on the original database.
-- Share-Alike: If you publicly use any adapted version of this database, or works produced from an adapted database, you must also offer that adapted database under the ODbL.
-- Keep open: If you redistribute the database, or an adapted version of it, then you may use technological measures that restrict the work (such as DRM) as long as you also redistribute a version without such measures.
+- **Attribute**: You must attribute any public use of the database, or works produced from the database, in the manner specified in the ODbL. For any use or redistribution of the database, or works produced from it, you must make clear to others the license of the database and keep intact any notices on the original database.
+- **Share-Alike**: If you publicly use any adapted version of this database, or works produced from an adapted database, you must also offer that adapted database under the ODbL.
+- **Keep open**: If you redistribute the database or an adapted version of it, then you may use technological measures that restrict the work (such as DRM) as long as you also redistribute a version without such measures.
 
 In short, publicly using OSM data without attribution or selling datasets derived from it is illegal.
 See the [License/Use Cases page on the OSM wiki](https://wiki.openstreetmap.org/wiki/License/Use_Cases) for detailed use cases.
 
 # Main package functions
 
-The packages is composed of the following main functions: 
+The package is composed of the following main functions: 
 
 1. `oe_providers()`: Show which providers of OSM extracts are available;
 1. `oe_match()`: Match one input place with one of the files stored by the OSM providers;
@@ -62,14 +62,14 @@ oe_providers()
 Each element in the column `database_name` is a data object that is packaged with `osmextract`. 
 You can read a detailed description of each provider data running, for example, `?geofabrik_zones` or `?bbbike_zones`. 
 
-Perhaps the best known bulk OSM data provider is Geofabrik, represented as data frame in the packaged object `geofabrik_zones`.
+Perhaps the best known bulk OSM data provider is Geofabrik, represented as `data.frame` in the packaged object `geofabrik_zones`.
 
 ```{r}
 class(geofabrik_zones)
 ```
 
 Note that in addition to being a data frame with rows and columns, the object is also an `sf` object, as defined in the [package](https://r-spatial.github.io/sf/) of the same name.
-When working with `sf` objects it makes sense to have the package loaded:
+When working with `sf` objects, it makes sense to have the package loaded:
 
 ```{r}
 library(sf)
@@ -84,11 +84,13 @@ plot(st_geometry(geofabrik_zones))
 ```
 
 The plot above shows how the provider divides geographic space into discrete chunks.
-Different providers have other zoning systems.
+Different providers have other zoning systems. 
+For example: 
 
 ```{r}
 par(mar = rep(0.1, 4))
-plot(st_geometry(bbbike_zones), xlim = c(-2, 10), ylim = c(30, 60))
+plot(st_geometry(spData::world), xlim = c(-2, 10), ylim = c(35, 60))
+plot(st_geometry(bbbike_zones), xlim = c(-2, 10), ylim = c(35, 60), add = TRUE, col = "darkred")
 ```
 
[…]
