+++ man/geofabrik_zones.Rd
@@ -4,7 +4,7 @@
 \name{geofabrik_zones}
 \alias{geofabrik_zones}
 \alias{test_zones}
-\title{A \code{data.frame} of geographical zones taken from Geofabrik}
+\title{An \code{sf} object of geographical zones taken from Geofabrik}
 \format{
 An sf object with 430 rows and
 15 columns:
