### 2020-09-28 38a029f
Explicit description of test_zones
