+++ man/geofabrik_zones.Rd
@@ -3,7 +3,6 @@
 \docType{data}
 \name{geofabrik_zones}
 \alias{geofabrik_zones}
-\alias{test_zones}
 \title{An \code{sf} object of geographical zones taken from Geofabrik}
 \format{
 An sf object with 430 rows and
