### 2020-09-28 f1b6e6d
@name instead of @rdname to prevent showing s3 methods in tab completion
