### 2020-09-28 41fcf09
minor tweaks to README
