### 2020-09-28 8d0263e
Add extra check to vectortranslate
