### 2020-09-28 f8417a6
test interaction between osmconf_ini and extra_tags
fix #117
