+++ man/oe_get.Rd
@@ -90,9 +90,10 @@ in the resulting data frame. See documentation at
 \href{https://gdal.org/drivers/vector/osm.html}{gdal.org}. Check Details in
 \code{\link[=oe_vectortranslate]{oe_vectortranslate()}}.}
 
-\item{extra_tags}{Which addition columns, corresponding to OSM tags,
-should be in the resulting dataset? \code{FALSE} by default. Check Details at
-\code{\link[=oe_vectortranslate]{oe_vectortranslate()}} and \code{\link[=oe_get_keys]{oe_get_keys()}} .}
+\item{extra_tags}{Which addition columns, corresponding to OSM tags, should
+be in the resulting dataset? \code{FALSE} by default. Check Details at
+\code{\link[=oe_vectortranslate]{oe_vectortranslate()}} and \code{\link[=oe_get_keys]{oe_get_keys()}}. Ignored when \code{osmconf_ini} is
+not \code{NULL}.}
 
 \item{force_vectortranslate}{Boolean. Force the original \code{.pbf} file to be
 translated into a \code{.gpkg} file, even if a \code{.gpkg} with the same name
