### 2020-09-28 43f3f38
Change default value of provider in oe_read
I think that we should have the same default values for provider argument between oe_read and oe_download
