### 2020-09-28 cbb9e64
Add more examples to oe_get
