### 2020-09-28 bbc6787
Second half of intro vignette
