+++ vignettes/osmextract.Rmd
@@ -46,7 +46,7 @@ The package is composed of the following main functions:
 1. `oe_download()`: Download the chosen file;
 1. `oe_vectortranslate()`: Convert between `.pbf` and `.gpkg` formats;
 1. `oe_read()`: Read `.pbf` and `.gpkg` files;
-1. `oe_get()`: Match, download, translate and import data, all in one step.
+1. `oe_get()`: Match, download, translate, and import data, all in one step.
 
 For many users who just want to get OSM data quickly `oe_get()` may be sufficient, as covered in the README.
 We will demonstrate each function in turn.
@@ -68,14 +68,14 @@ Perhaps the best known bulk OSM data provider is Geofabrik, represented as `data
 class(geofabrik_zones)
 ```
 
-Note that in addition to being a data frame with rows and columns, the object is also an `sf` object, as defined in the [package](https://r-spatial.github.io/sf/) of the same name.
+Note that in addition to being a data frame with rows and columns, `geofabrik_zones` is also an `sf` object, as defined in the [package](https://r-spatial.github.io/sf/) of the same name.
 When working with `sf` objects, it makes sense to have the package loaded:
 
 ```{r}
 library(sf)
 ```
 
-That gives you access to many geographic functions for working with geographic vector data of the type provided by `osmextract`.
+That gives you access to many functions for working with geographic vector data of the type provided by `osmextract`.
 Each row of data in an `sf` object contains a geometry, representing the area covered by each provider zone, meaning you can plot the data as follows:
 
 ```{r}
@@ -87,14 +87,18 @@ The plot above shows how the provider divides geographic space into discrete chu
 Different providers have other zoning systems. 
 For example: 
 
-```{r}
+```{r, eval = FALSE}
 par(mar = rep(0.1, 4))
 plot(st_geometry(spData::world), xlim = c(-2, 10), ylim = c(35, 60))
-plot(st_geometry(bbbike_zones), xlim = c(-2, 10), ylim = c(35, 60), add = TRUE, col = "darkred")
+plot(st_geometry(bbbike_zones), xlim = c(-2, 10), ylim = c(35, 60), col = "darkred", add = TRUE)
 ```
 
+<img src="https://user-images.githubusercontent.com/22221146/94461461-772e4d00-01ba-11eb-950c-804ad177729f.png" width="100%" />
+
 As shown in the above visualisation of [BBBike.org](https://download.bbbike.org/osm/) zones in Europe, the provider offers rectangular extracts of major cities.
-You can also download manually selected regions of interest from the BBBike website (see also https://github.com/ITSLeeds/osmextract/issues/100).
+We are working on adding support for manually selected regions from the BBBike website (see https://github.com/ITSLeeds/osmextract/issues/100).
+
+Check **ADD LINK** for a more detailed comparison between the supported providers. 
 
 ## `oe_match()`: finding OSM extracts
 
@@ -115,7 +119,7 @@ Nevertheless, there are several situations where it could be challenging to find
 oe_match("Russia", quiet = FALSE)
 ```
 
-Therefore, you can select alternative fields in the provider's data (instead of the default field, i.e. `name`), that can be specified using the `match_by` parameter. 
+Therefore, you can select alternative fields in the provider's data (instead of the default field, i.e. `name`) using the `match_by` parameter. 
 For example, if you work with `geofabrik_data`, you can look for a match using [iso3166-1 alpha2](https://it.wikipedia.org/wiki/ISO_3166-1_alpha-2) codes: 
 
 ```{r}
@@ -130,19 +134,34 @@ oe_match("PS", match_by = "iso3166_1_alpha2")
 oe_match("IL", match_by = "iso3166_1_alpha2")
 ```
 
[…]
