### 2020-09-29 6d2a73b
minor tweaks to new provider vignette
