### 2020-09-29 17c9f71
explicit op operation
