### 2020-09-29 66e31c4
Add new vignette
Fix #127
