### 2020-09-30 a25ec47
Fix minor typo in docs
