+++ pkgdown.yml
@@ -4,5 +4,6 @@ pkgdown_sha: e1e46f2e13c7efdb17bc4eae40bc3779c2b0b517
 articles:
   osmextract: osmextract.html
   providers: providers.html
-last_built: 2020-09-23T05:52Z
+  providers_comparisons: providers_comparisons.html
+last_built: 2020-09-30T08:15Z
 
