### 2020-10-01 1655ab4
Add warning for version of GDAL that do not support WKT.
Fix #132
