+++ pkgdown.yml
@@ -5,5 +5,5 @@ articles:
   osmextract: osmextract.html
   providers: providers.html
   providers_comparisons: providers_comparisons.html
-last_built: 2020-09-30T08:15Z
+last_built: 2020-10-01T06:38Z
 
