### 2020-10-01 #395 @agila5
Hi @annakrystalli. First of all, thank you very much for testing our R package and helping us make it better.

> I've now completed an initial pass of editor's checks. There's a few minor points but most importantly there is n error when building one of the vignettes (see below) and this is causing problems for me to fully complete the checks.

I tried to reproduce the errors using my laptop (running Windows 10) and a server (running Ubuntu 16.04), but I couldn't replicate it. Could you please share the output of `sessionInfo()` and `sf::sf_extSoftVersion()`? @Robinlovelace suggested it may be a problem related to an old version of GDAL. Moreover, we recently changed the `osmextract.Rmd` vignette, could you please try running the checks using the most recent version of the package?

I think that the error in `devtools::test` has the same root cause as the error when installing the package.

> and bunch of warnings

The warning messages are caused by a version of GDAL that does not support WKT, while the provider's data in `osmextract` was created using WKT. The problem is documented in the [sf issue tracker](https://github.com/r-spatial/sf/issues/1419). I added a note in the README:

https://github.com/ITSLeeds/osmextract/blob/1655ab4186c9a1402584234c3a9e808659940044/README.Rmd#L178-L187

and a warning message in `.onLoad()`:

https://github.com/ITSLeeds/osmextract/blob/1655ab4186c9a1402584234c3a9e808659940044/R/zzz.R#L12-L25

These warnings will tend to fade as OSs update and should disapper when installing newer version of PROJ and GDAL.

Sorry for these problems, we'll try to fix them as soon as possible.
