### 2020-10-01 #395 @mpadge
> These warnings will tend to fade as OSs update and should disapper when installing newer version of PROJ and GDAL.

For what it's worth, the package installs and tests clean on my (linux) system with absolute latest versions of everything (well, okay, not yet R4.0.3 :smirk:):

    R.Version()$version.string
    #> [1] "R version 4.0.2 (2020-06-22)"
    system2 ("gdalinfo", "--version", stdout = TRUE)
    #> [1] "GDAL 3.0.4, released 2020/01/28"
    system2 ("proj", "--version", stdout = TRUE)
    #> [1] "Rel. 6.3.2, May 1st, 2020"

(Tests just give the couple of warnings about `st_centroid` applied to lon-lat data.)
