### 2020-10-02 b94cebf
Add URL check in oe_read
Fix #134
