### 2020-10-02 d0b330b
abstract URL regex into a new function
