+++ R/read.R
@@ -86,7 +86,7 @@ oe_read = function(
     # Add an if clause to check if file_path "looks like" a URL
     # See https://github.com/ITSLeeds/osmextract/issues/134 and
     # https://stackoverflow.com/questions/3809401/what-is-a-good-regular-expression-to-match-a-url
-    like_url <- is_like_url(file_path)
+    like_url = is_like_url(file_path)
 
     if (!like_url) {
       stop(
