### 2020-10-02 e26c18b
Update utils.R
