+++ R/utils.R
@@ -2,7 +2,7 @@
 '%!in%' = function(x, y) !('%in%'(x,y))
 
 # See https://github.com/ITSLeeds/osmextract/issues/134
-is_like_url <- function(URL) {
+is_like_url = function(URL) {
   grepl(
     pattern = "https?:\\/\\/(www\\.)?[-a-zA-Z0-9@:%._\\+~#=]{1,256}\\.[a-zA-Z0-9()]{1,6}\\b([-a-zA-Z0-9()@:%_\\+.~#?&//=]*)",
     x = URL,
