+++ pkgdown.yml
@@ -5,5 +5,5 @@ articles:
   osmextract: osmextract.html
   providers: providers.html
   providers_comparisons: providers_comparisons.html
-last_built: 2020-10-01T06:38Z
+last_built: 2020-10-02T14:42Z
 
