### 2020-10-04 #395 @annakrystalli
Thanks for the prompt response.  The message is useful as is the note in the README. However I would move that to the installation section and be explicit about the minimum versions of the GDAL and PROJ libraries required for full functionality.

Now onto the fix I had a few issues with (so anticipate others might too!). Apologies for not sharing platform details earlier! Indeed I am on macos, here's full platform info:

``` r
sessioninfo::platform_info()
#>  setting  value
#>  version  R version 3.6.2 (2019-12-12)
#>  os       macOS Catalina 10.15.6
#>  system   x86_64, darwin15.6.0
#>  ui       X11
#>  language (EN)
#>  collate  en_GB.UTF-8
#>  ctype    en_GB.UTF-8
#>  tz       Europe/London
#>  date     2020-10-04
```

I've now updated `GDAL` to 3.1.2 using homebrew which also updated `PROJ4` library to 7.1.1

```
==> Upgrading 1 outdated package:
gdal 2.4.4_2 -> 3.1.2_1

==> Installing gdal dependency: proj
==> Pouring proj-7.1.1.catalina.bottle.tar.gz
```

Just a note that I kept getting an error even after reinstalling `sf` from CRAN. It was still linking to an earlier version of proj.

```r
> library(sf)
Linking to GEOS 3.7.2, GDAL 2.4.2, PROJ 5.2.0
```

FYI, I solved it by:
- Set environment variable `PKG_CONFIG_PATH` to, in my case, `/usr/local/lib/pkgconfig/`, the path to to the `pkgconfig/` directory containing a `proj.pc` file with details of the newest version, as suggested in this [issue](https://github.com/r-spatial/sf/issues/335).
- Reinstalled `sf` from GitHub with `remotes::install_github("r-spatial/sf")` (reinstalling from CRAN didn't solve the linking problem)

Which now works!
``` r
library("sf")
#> Linking to GEOS 3.8.1, GDAL 3.1.2, PROJ 7.1.1
```
But! I'm still having issues, see below:

### tests

```
Loading osmextract
Data (c) OpenStreetMap contributors, ODbL 1.0. https://www.openstreetmap.org/copyright.
Any product made from OpenStreetMap must cite OSM as the data source.
Geofabrik data are taken from https://download.geofabrik.de/
For usage details of bbbike data see https://download.bbbike.org/osm/
OpenStreetMap_fr data are taken from http://download.openstreetmap.fr/
Testing osmextract
✓ |  OK F W S | Context
⠏ |   0       | downloadtrying URL 'https://github.com/ITSLeeds/osmextract/raw/master/inst/its-example.osm.pbf'
Content type 'application/octet-stream' length 40792 bytes (39 KB)
==================================================
downloaded 39 KB

✓ |   6       | download [1.0 s]
✓ |   3       | find [0.6 s]
✓ |   1       | get
✓ |  26   3   | match [0.7 s]
───────────────────────────────────────────────────────────────────────────────
test-match.R:38: warning: oe_match: sfc_POINT objects
st_centroid does not give correct centroids for longitude/latitude data

test-match.R:43: warning: oe_match: sfc_POINT objects
st_centroid does not give correct centroids for longitude/latitude data

test-match.R:43: warning: oe_match: sfc_POINT objects
st_centroid does not give correct centroids for longitude/latitude data
───────────────────────────────────────────────────────────────────────────────
✓ |   1       | providers
x |   1 1   1 | read
───────────────────────────────────────────────────────────────────────────────
test-read.R:3: error: oe_read: simplest examples work
The input file_path does not correspond to any existing file and it doesn't look like a URL.
Backtrace:
 1. osmextract::oe_read(f) tests/testthat/test-read.R:3:2

test-read.R:23: skip: or_read: simplest example with a URL works
Reason: empty test
───────────────────────────────────────────────────────────────────────────────
⠏ |   0       | updateThe .gpkg files are going to be removed.
✓ |   1       | update [0.6 s]
✓ |   7       | vectortranslate [0.3 s]

══ Results ════════════════════════════════════════════════════════════════════
Duration: 3.4 s

OK:       46
Failed:   1
Warnings: 3
Skipped:  1
```

### checks

Also a couple more related errors in examples being thrown up by `devtools::check()`

```r
E  checking examples (2.8s)
   Running examples in ‘osmextract-Ex.R’ failed
   The error most likely occurred in:

   > base::assign(".ptime", proc.time(), pos = "CheckExEnv")
   > ### Name: oe_read
   > ### Title: Read a '.pbf' or '.gpkg' object from file or URL
   > ### Aliases: oe_read
   >
   > ### ** Examples
   >
   > # Read an existing .pbf file
   > my_pbf = system.file("its-example.osm.pbf", package = "osmextract")
   > oe_read(my_pbf, quiet = FALSE)
   Error: The input file_path does not correspond to any existing file and it doesn't look like a URL.
   Execution halted
✓  checking for unstated dependencies in ‘tests’ ...
─  checking tests ...
E  Running ‘testthat.R’ (4s)
   Last 13 lines of output:
     ==================================================
     downloaded 39 KB

     ── 1. Error: oe_read: simplest examples work (@test-read.R
     The input file_path does not correspond to any existing file and it doesn't look like a URL.
     Backtrace:
      1. osmextract::oe_read(f)

     The .gpkg files are going to be removed.
     ══ testthat results  ═════════════════════════════════════
     [ OK: 46 | SKIPPED: 1 | WARNINGS: 3 | FAILED: 1 ]
     1. Error: oe_read: simplest examples work (@test-read.R#3)

     Error: testthat unit tests failed
     Execution halted
✓  checking for unstated dependencies in vignettes ...
✓  checking package vignettes in ‘inst/doc’ ...
✓  checking re-building of vignette outputs (10.6s)
✓  checking for detritus in the temp directory

   See
     ‘/Users/Anna/Documents/workflows/rOpenSci/editorials/osmextract.Rcheck/00check.log’
   for details.

── R CMD check results ─────────────────────────────────── osmextract 0.1.0 ────
Duration: 46.5s

> checking examples ... ERROR
  Running examples in ‘osmextract-Ex.R’ failed
  The error most likely occurred in:

  > base::assign(".ptime", proc.time(), pos = "CheckExEnv")
  > ### Name: oe_read
  > ### Title: Read a '.pbf' or '.gpkg' object from file or URL
  > ### Aliases: oe_read
  >
  > ### ** Examples
  >
  > # Read an existing .pbf file
  > my_pbf = system.file("its-example.osm.pbf", package = "osmextract")
  > oe_read(my_pbf, quiet = FALSE)
  Error: The input file_path does not correspond to any existing file and it doesn't look like a URL.
  Execution halted
```

### goodpractice

A few more minor issues thrown up by GP

```r
── GP osmextract ──────────────────────────────────────────────────────────────

It is good practice to

  ✖ use '<-' for assignment instead of '='. '<-' is the
    standard, and R users and developers are used it and it is easier
    to read your code for them if you use '<-'.

    R/download.R:61:13
    R/download.R:83:14
    R/download.R:88:13
    R/download.R:110:14
    R/download.R:119:16
    ... and 122 more lines

    R/download.R:40:1
    R/find.R:23:1
    R/find.R:57:1
    R/get.R:9:1
    R/get.R:73:1
    ... and 22 more lines

  ✖ fix this R CMD check NOTE: Note: found 263 marked UTF-8
    strings
```

### spellcheck

One valid (from what I can tell) spelling error thrown up by `devtools::spell_check()`

```
Antartica             geofabrik_zones.Rd:18

```

### Coverage

I still can't get a coverage breakdown because I'm getting the same error when running `covr::package_coverage()`

```
Error: Failure in `/private/var/folders/8p/87cqdx2s34vfvcgh04l6z72w0000gn/T/RtmppGl6ac/R_LIBS61956a128da7/osmextract/osmextract-tests/testthat.Rout.fail` .osm.pbf' Content type 'application/octet-stream' length 40792 bytes (39 KB) ================================================== downloaded 39 KB ── 1. Error: oe_read: simplest examples work (@test-read.R#3) ──────────────── The input file_path does not correspond to any existing file and it doesn't look like a URL. Backtrace: 1. osmextract::oe_read(f) The .gpkg files are going to be removed. ══ testthat results ══════════════════════════════════════════════════════════ [ OK: 46 | SKIPPED: 1 | WARNINGS: 3 | FAILED: 1 ] 1. Error: oe_read: simplest examples work (@test-read.R#3) Error: testthat unit tests failed Execution halted
```

Any ideas? It seems @mpadge didn't hit this issue 🤔
