### 2020-10-06 03fe3d4
move warning messages to installation section and set explicit minimal requirements
skip-ci
