+++ man/oe_get.Rd
@@ -174,7 +174,8 @@ baku = oe_get(place = "Baku", provider = "bbbike", quiet = FALSE)
 
 # Other examples:
 oe_get("RU", match_by = "iso3166_1_alpha2", quiet = FALSE)
-oe_get("Andora", stringsAsFactors = FALSE, quiet = TRUE, as_tibble = TRUE) # like read_sf
+# The following example mimics read_sf
+oe_get("Andora", stringsAsFactors = FALSE, quiet = TRUE, as_tibble = TRUE)
 }
 }
 \seealso{
