### 2020-10-06 #395 @agila5
Hi @annakrystalli, and thank you very much again for your feedback.

> The message is useful as is the note in the README. However I would move that to the installation section and be explicit about the minimum versions of the GDAL and PROJ libraries required for full functionality.

I just moved the warning message in the installation section, thanks!

> goodpractice
> A few more minor issues thrown up by GP

We decided to adopt `=` as the assignment operator, so the first set of warning messages is somewhat "expected". I recently changed some of the "long lines of code" and, at the moment, the remaining ones should represent only long URL(s). I'm not sure about the UTF8 strings, but I think that's not a big deal.

> spellcheck
> One valid (from what I can tell) spelling error thrown up by devtools::spell_check()

Fixed!

> Now onto the fix I had a few issues with (so anticipate others might too!). Apologies for not sharing platform details earlier! Indeed I am on macos, here's full platform info:

First of all, thanks for your useful feedbacks , I will add a link to the README pointing to this discussion as soon as we fix these problems. Unfortunately, I have very little experience with MacOS + `sf`, so I have no idea what's going on 😢 The only thing that I may suggest is running (I read the following command [here](https://github.com/r-spatial/sf/issues/1502#issuecomment-702823684)):

``` r
install.packages("sf", type = "mac.binary")
remotes::install_github("ITSLeeds/osmextract")
```

The first line of code should overwrite the previous version of `sf`, and install a "new" binary version with its own GEOS/GDAL/PROJ. The main drawback is that this "new" version of `sf` is not linked to the external GEOS/GDAL/PROJ (and this is the same behaviour that occurr on Windows, I think. @RobinLovelace please correct me here). I checked this approach with a friend of mine, and we installed the `osmextract` package without any problem.

Another question: do you face any problem when you install another R package that depends on `sf`?
