### 2020-10-07 #395 @annakrystalli
Hi @agila5 ! And thanks for your response!

BTW I did manage to install! I just had to follow these steps to get everything to work properly:

> FYI, I solved it by:
> - Set environment variable PKG_CONFIG_PATH to, in my case, /usr/local/lib/pkgconfig/, the path to to the pkgconfig/ directory containing a proj.pc file with details of the newest version, as suggested in this issue.
> - Reinstalled sf from GitHub with remotes::install_github("r-spatial/sf") (reinstalling from CRAN didn't solve the linking problem)
> Which now works!

But do you think the test failure and broken example still are an `sf` installation issue? The main issue in tests and checks seems to be:
```r
  > my_pbf = system.file("its-example.osm.pbf", package = "osmextract")
  > oe_read(my_pbf, quiet = FALSE)
  Error: The input file_path does not correspond to any existing file and it doesn't look like a URL.
```
