### 2020-10-07 #395 @agila5
> But do you think the test failure and broken example still are an sf installation issue?

I think that the error is caused by some problems related to the installation of `osmextract`. The function `oe_read` is used to read-in a `.pbf` or `.gpkg` file, that can be specified using a file-path or a URL. It should fail only in two cases: 1) the input data points to a non-existing file or 2) the input data points to a non-existing URL.

The object `my_pbf` should represent the path to the `its-example.osm.pbf` file, which is one of the files included in the package (see https://github.com/ITSLeeds/osmextract/tree/master/inst). For example, I see:

``` r
list.files("C:/Users/Utente/Documents/R/win-library/3.6/osmextract/")
#>  [1] "data"                "DESCRIPTION"         "help"
#>  [4] "html"                "INDEX"               "its-example.osm.pbf"
#>  [7] "Meta"                "NAMESPACE"           "osmconf.ini"
#> [10] "R"
```

Could you compare that output with the analogous command on macOS?  That example is clearly failing, and (I think) that should imply that there were some problems during the installation of the package. Could you try reinstalling it again?

I can succesfully replicate the same behaviour only when I run `devtools::test` after deleting the `"its-example.osm.pbf"` file (which is clearly a patological example):

``` r
# remove "old" osmextract
remove.packages("osmextract")
#> Removing package from 'C:/Users/Utente/Documents/R/win-library/3.6'
#> (as 'lib' is unspecified)

# check that it's removed
library("osmextract")
#> Error in library("osmextract"): there is no package called 'osmextract'

# download master branch
osmextract_master <- tempfile(fileext = ".zip")
download.file(
  url = "https://github.com/ITSLeeds/osmextract/archive/master.zip",
  destfile = osmextract_master
)

# unzip
unzip(osmextract_master, exdir = tempdir())
list.files(tempdir())
#> [1] "file2a2c23283a3e"     "file2a2c3f246c27"     "file2a2c54c36595.dll"
#> [4] "file2a2c7f3d50ae.zip" "osmextract-master"

# delete "its-example.osm.pbf" from osmextract-master
file.remove(file.path(tempdir(), "osmextract-master", "inst", "its-example.osm.pbf"))
#> [1] TRUE

# run tests
devtools::test(file.path(tempdir(), "osmextract-master"))
#> Loading osmextract
#> Data (c) OpenStreetMap contributors, ODbL 1.0. https://www.openstreetmap.org/copyright.
#> Any product made from OpenStreetMap must cite OSM as the data source.
#> Geofabrik data are taken from https://download.geofabrik.de/
#> For usage details of bbbike data see https://download.bbbike.org/osm/
#> OpenStreetMap_fr data are taken from http://download.openstreetmap.fr/
#> Testing osmextract
#> v |  OK F W S | Context
#> / |   0       | download- |   1       | downloadv |   6       | download [0.6 s]
#> / |   0       | find\ |   2       | findv |   3       | find [0.3 s]
#> / |   0       | getv |   1       | get
#> / |   0       | match/ |   4       | match\ |   6       | match- |  10   3   | matchv |  26   3   | match [0.5 s]
#> ----------------------------------------------------------------------------------------------------------------------------
#> test-match.R:38: warning: oe_match: sfc_POINT objects
#> st_centroid does not give correct centroids for longitude/latitude data
#>
#> test-match.R:43: warning: oe_match: sfc_POINT objects
#> st_centroid does not give correct centroids for longitude/latitude data
#>
#> test-match.R:43: warning: oe_match: sfc_POINT objects
#> st_centroid does not give correct centroids for longitude/latitude data
#> ----------------------------------------------------------------------------------------------------------------------------
#> / |   0       | providersv |   1       | providers
#> / |   0       | readx |   1 1   1 | read
#> ----------------------------------------------------------------------------------------------------------------------------
#> test-read.R:3: error: oe_read: simplest examples work
#> The input file_path does not correspond to any existing file and it doesn't look like a URL.
#> Backtrace:
#>  1. osmextract::oe_read(f) C:\Users\Utente\AppData\Local\Temp\Rtmp82J8ZK\osmextract-master/tests/testthat/test-read.R:3:2
#>
#> test-read.R:23: skip: or_read: simplest example with a URL works
#> Reason: empty test
#> ----------------------------------------------------------------------------------------------------------------------------
#> / |   0       | updateThe .gpkg files are going to be removed.
#> v |   1       | update
#> / |   0       | vectortranslate- |   1       | vectortranslate| |   3       | vectortranslatev |   7       | vectortranslate [0.4 s]
#>
#> == Results =================================================================================================================
#> Duration: 2.1 s
#>
#> OK:       46
#> Failed:   1
#> Warnings: 3
#> Skipped:  1
```

Sorry for all installation problems 😢
