### 2020-10-20 8156a10
Update .onLoad with old GDAL
Probably I shouldnt use warning in .onLoad message
