### 2020-10-21 76763b1
put everything in .onAttach
