+++ R/zzz.R
@@ -7,19 +7,18 @@
     "OpenStreetMap_fr data are taken from http://download.openstreetmap.fr/",
     sep = "\n"
   ))
-}
 
-.onLoad = function(libname, pkgname) {
   if (
     sf::sf_extSoftVersion()["GDAL"] < "3.0.0" ||
     sf::sf_extSoftVersion()["proj.4"] < "6.0.0"
   ) {
-    message(
+    packageStartupMessage(paste0(
       "The package may return several warning messages like\n",
       "'st_crs<- : replacing crs does not reproject data; use st_transform for that'\n",
       "They are caused by a version of GDAL that does not support WKT.",
       " The functions defined in this package should still work.",
       " Check the README for more details."
-    )
+    ))
   }
 }
+
