### 2020-10-21 19b8a39
Add checks for ... and st_read
should fix #114
