+++ NAMESPACE
@@ -1,5 +1,8 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(oe_get_keys,character)
+S3method(oe_get_keys,default)
+S3method(oe_get_keys,sf)
 S3method(oe_match,character)
 S3method(oe_match,default)
 S3method(oe_match,numeric)
