### 2020-11-02 #395 @annakrystalli
Success! I recloned the original source code and everything seemed to work fine now. Really sorry for the hold up!

I'll start looking for reviewers. In the meantime, for completeness, I've now also ran `package_coverage()` and everything seems good apart from `zzz.R` and `update.R` which are below ~75% coverage. Any particular reason for that?

``` r
pkg_dir <- "/Users/Anna/Documents/workflows/rOpenSci/editorials/osmextract"
covr::package_coverage(pkg_dir)
#> osmextract Coverage: 82.76%
#> R/zzz.R: 58.82%
#> R/update.R: 68.25%
#> R/read.R: 74.19%
#> R/vectortranslate.R: 79.43%
#> R/download.R: 82.69%
#> R/find.R: 86.84%
#> R/match.R: 88.73%
#> R/utils.R: 91.67%
#> R/get.R: 100.00%
#> R/providers.R: 100.00%
```
