### 2020-11-02 #395 @agila5
Hi @annakrystalli and thanks for your comments. I'm really happy that now it's working fine 🎉 🎉 🎉 I will add a link in the README that points to this discussion since it could help new users.

> In the meantime, for completeness, I've now also ran package_coverage() and everything seems good apart from zzz.R and update.R which are below ~75% coverage. Any particular reason for that?

The file `zzz.R` includes only `.onAttach()` and the code lines that are not covered by tests are used to print an (hopefully) informative message in case GDAL or PROJ are "too old". I think that I cannot add any significative test in this case. Tomorrow I will add more tests for `read.R` and `update.R`!
