### 2020-11-03 9aa1229
Add more tests
See https://github.com/ropensci/software-review/issues/395
