### 2020-11-04 b6b5218
Describe how to workaround downloads timing out. #143
