+++ README.Rmd
@@ -187,6 +187,10 @@ You can always check the default `download_directory` used by `oe_get()` with:
 oe_download_directory()
 ```
 
+## Troubleshooting
+
+Depending on the `.pbf` file selected and your connection speed, you may experience an error stating `Timeout of 60 seconds was reached`. If so, before calling `oe_get()`, you can adjust the timeout using `options(timeout = 300)`, choosing an appropriate value. This setting affects all calls to [download.file](https://www.rdocumentation.org/packages/utils/versions/3.6.2/topics/download.file), so you may need to reset it for the rest of your script.
+
 ## Next steps
 
 We hope to make the user interface to the SQL syntax more user friendly.
