+++ README.md
@@ -1,42 +1,18 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-
 # osmextract
 
 <!-- badges: start -->
+[![R build status](https://github.com/itsleeds/osmextract/workflows/R-CMD-check/badge.svg)](https://github.com/itsleeds/osmextract/actions) [![Codecov test coverage](https://codecov.io/gh/itsleeds/osmextract/branch/master/graph/badge.svg)](https://codecov.io/gh/itsleeds/osmextract?branch=master) <!-- badges: end -->
 
-[![R build
-status](https://github.com/itsleeds/osmextract/workflows/R-CMD-check/badge.svg)](https://github.com/itsleeds/osmextract/actions)
-[![Codecov test
-coverage](https://codecov.io/gh/itsleeds/osmextract/branch/master/graph/badge.svg)](https://codecov.io/gh/itsleeds/osmextract?branch=master)
-<!-- badges: end -->
+The goal of `osmextract` is to make it easier for people to access OpenStreetMap (OSM) data for reproducible research. OSM data is the premier source of freely available, community created geographic data worldwide. We aim to enable you to extract it for data-driven work in the public interest.
 
-The goal of `osmextract` is to make it easier for people to access
-OpenStreetMap (OSM) data for reproducible research. OSM data is the
-premier source of freely available, community created geographic data
-worldwide. We aim to enable you to extract it for data-driven work in
-the public interest.
-
-`osmextract` downloads, converts and imports bulk OSM data hosted by
-providers such as [Geofabrik GmbH](http://download.geofabrik.de) and
-[bbbike](https://download.bbbike.org/osm/). For information on
-alternatives and how to add them see the [providers
-vignette](https://itsleeds.github.io/osmextract/articles/providers.html).
+`osmextract` downloads, converts and imports bulk OSM data hosted by providers such as [Geofabrik GmbH](http://download.geofabrik.de) and [bbbike](https://download.bbbike.org/osm/). For information on alternatives and how to add them see the [providers vignette](https://itsleeds.github.io/osmextract/articles/providers.html).
 
 ## Why osmextract?
 
-The package answers a common question for researchers who use OSM data:
-how to get it into a statistical environment, in an appropriate format,
-as part of a computationally efficient and reproducible workflow? Other
-packages answer parts of this question.
-[`osmdata`](https://github.com/ropensci/osmdata), for example, is an R
-package that provides an R interface to the [Overpass
-API](https://wiki.openstreetmap.org/wiki/Overpass_API), which is ideal
-for downloading small OSM datasets. However, the API is rate limited,
-making it hard to download large datasets. As a case study, try to
-download all cycleways in England using `osmdata`:
+The package answers a common question for researchers who use OSM data: how to get it into a statistical environment, in an appropriate format, as part of a computationally efficient and reproducible workflow? Other packages answer parts of this question. [`osmdata`](https://github.com/ropensci/osmdata), for example, is an R package that provides an R interface to the [Overpass API](https://wiki.openstreetmap.org/wiki/Overpass_API), which is ideal for downloading small OSM datasets. However, the API is rate limited, making it hard to download large datasets. As a case study, try to download all cycleways in England using `osmdata`:
 
 ``` r
 library(osmdata)
@@ -47,14 +23,7 @@ cycleways_england = opq("England") %>%
 # The data included in this document is from www.openstreetmap.org. The data is made available under ODbL. runtime error: Query timed out in "query" at line 4 after 26 seconds. 
 ```
 
[…]
