### 2020-11-05 e3e438b
Update CI - faster runs, simpler set-up
