+++ .github/workflows/R-CMD-check.yaml
@@ -12,9 +12,10 @@ on:
 
 name: R-CMD-check
 
+
 jobs:
   R-CMD-check:
-    runs-on: ${{ matrix.config.os }}
+    if: "!(contains(github.event.head_commit.message, 'skip ci')||contains(github.event.head_commit.message, 'ci skip')||contains(github.event.head_commit.message, 'skip-ci')||contains(github.event.head_commit.message, 'ci-skip')||contains(github.event.head_commit.message, 'skip_ci')||contains(github.event.head_commit.message, 'ci_skip'))"
 
     name: ${{ matrix.config.os }} (${{ matrix.config.r }})
 
