+++ pkgdown.yml
@@ -1,9 +1,9 @@
 pandoc: 2.7.3
 pkgdown: 1.6.1.9000
-pkgdown_sha: e5847476d79f08b4cdca0ed5cc55bf55fc18a161
+pkgdown_sha: aaf1509f3c9f4f3677ee01e11299dff4a0d696fb
 articles:
   osmextract: osmextract.html
   providers: providers.html
   providers_comparisons: providers_comparisons.html
-last_built: 2020-11-04T20:09Z
+last_built: 2020-11-05T11:49Z
 
