+++ R/zzz.R
@@ -22,3 +22,5 @@
   }
 }
 
+
+# :)
