### 2020-11-08 b89df15
minor tweaks
