### 2020-11-08 b88a202
Add a new parameter called never_skip_vectortranslate
Should fix #144
