### 2020-11-10 #395 @annakrystalli
One last thing @agila5 , could you please add the rOpenSci under review badge to your README?
I'll be back in touch when I have found both reviewers 👍

---

Reviewers:
Due date:
