+++ README.Rmd
@@ -21,7 +21,7 @@ knitr::opts_chunk$set(
 
 [![R build status](https://github.com/itsleeds/osmextract/workflows/R-CMD-check/badge.svg)](https://github.com/itsleeds/osmextract/actions)
 [![Codecov test coverage](https://codecov.io/gh/itsleeds/osmextract/branch/master/graph/badge.svg)](https://codecov.io/gh/itsleeds/osmextract?branch=master)
-[![peer-review](https://badges.ropensci.org/395_status.svg)](https://github.com/ropensci/software-review/issues/395)
+[![](https://badges.ropensci.org/395_status.svg)](https://github.com/ropensci/software-review/issues/395)
 <!-- badges: end -->
 
 The goal of `osmextract` is to make it easier for people to access OpenStreetMap (OSM) data for reproducible research.
@@ -39,6 +39,8 @@ Other packages answer parts of this question.
 [`osmdata`](https://github.com/ropensci/osmdata), for example, is an R package that provides an R interface to the [Overpass API](https://wiki.openstreetmap.org/wiki/Overpass_API), which is ideal for downloading small OSM datasets.
 However, the API is rate limited, making it hard to download large datasets.
 As a case study, try to download all cycleways in England using `osmdata`:
+
+![peer-review](https://badges.ropensci.org/395_status.svg)
 
 ```r
 library(osmdata)
