+++ README.md
@@ -11,7 +11,7 @@
 status](https://github.com/itsleeds/osmextract/workflows/R-CMD-check/badge.svg)](https://github.com/itsleeds/osmextract/actions)
 [![Codecov test
 coverage](https://codecov.io/gh/itsleeds/osmextract/branch/master/graph/badge.svg)](https://codecov.io/gh/itsleeds/osmextract?branch=master)
-[![peer-review](https://badges.ropensci.org/395_status.svg)](https://github.com/ropensci/software-review/issues/395)
+[![](https://badges.ropensci.org/395_status.svg)](https://github.com/ropensci/software-review/issues/395)
 <!-- badges: end -->
 
 The goal of `osmextract` is to make it easier for people to access
@@ -38,6 +38,8 @@ API](https://wiki.openstreetmap.org/wiki/Overpass_API), which is ideal
 for downloading small OSM datasets. However, the API is rate limited,
 making it hard to download large datasets. As a case study, try to
 download all cycleways in England using `osmdata`:
+
+![peer-review](https://badges.ropensci.org/395_status.svg)
 
 ``` r
 library(osmdata)
