### 2020-11-11 #395 @agila5
Hi @annakrystalli and thank you very much for your assistance.

> One last thing @agila5 , could you please add the rOpenSci under review badge to your README?

I added the badge to the `README.Rmd` but when I knit the document I see the following error:

```
HttpExceptionRequest Request {
  port                 = 443
  secure               = True
  requestHeaders       = []
  queryString          = ""
  method               = "GET"
  proxy                = Nothing
  rawBody              = False
  redirectCount        = 10
  responseTimeout      = ResponseTimeoutDefault
  requestVersion       = HTTP/1.1
}
Errore: pandoc document conversion failed with error 61
Esecuzione interrotta
```

I cannot see the badge on `README.md` but it's working on the website generated by `pkgdown`
