+++ README.Rmd
@@ -54,11 +54,9 @@ The same query can be made with `osmextract` as follows, which reads-in almost 1
 The download-conversion operation of the OSM extract associated to England takes approximately a few minutes, and this operation must be executed only once. 
 The following code chunk is not evaluated:
 
-```{r}
+```{r, eval = FALSE}
 library(osmextract)
-```
 
-```{r, eval=FALSE}
 cycleways_england = oe_get(
   "England",
   quiet = FALSE,
@@ -159,13 +157,13 @@ osm_major_roads = osm_lines[osm_lines$highway %in% ht, ]
 plot(osm_major_roads["highway"], key.pos = 1)
 ```
 
-The same steps can be used to get other OSM datasets (note use of `quiet = FALSE` to show additional message, examples not run):
+The same steps can be used to get other OSM datasets (examples not run):
 
 ```{r, eval = FALSE}
-test_malta = oe_get("Malta", quiet = FALSE)
-test_andorra = oe_get("Andorra", extra_tags = "ref", quiet = FALSE)
-test_leeds <- oe_get("Leeds", provider = "bbbike", quiet = FALSE)
-test_india_region <- oe_get("Goa", provider = "openstreetmap_fr", quiet = FALSE)
+test_malta = oe_get("Malta", quiet = TRUE)
+test_andorra = oe_get("Andorra", extra_tags = "ref")
+test_leeds <- oe_get("Leeds", provider = "bbbike")
+test_india_region <- oe_get("Goa", provider = "openstreetmap_fr")
 ```
 
 For further details on using the package, see the [Introducing osmextract vignette](https://itsleeds.github.io/osmextract/articles/osmextract.html).
@@ -190,7 +188,7 @@ oe_download_directory()
 
 ## Troubleshooting
 
-Depending on the `.pbf` file selected and your connection speed, you may experience an error stating `Timeout of 60 seconds was reached`. If so, before calling `oe_get()`, you can adjust the timeout using `options(timeout = 300)`, choosing an appropriate value. This setting affects all calls to [download.file](https://www.rdocumentation.org/packages/utils/versions/3.6.2/topics/download.file), so you may need to reset it for the rest of your script.
+Depending on the `.pbf` file selected and your connection speed, you may experience an error stating `Timeout of 60 seconds was reached`. If so, before calling `oe_get()`, you can adjust the timeout using `options(timeout = 300)`, choosing an appropriate value. This setting affects all calls to [download.file()](https://www.rdocumentation.org/packages/utils/versions/3.6.2/topics/download.file), so you may need to reset it for the rest of your script.
 
 ## Next steps
 
