### 2020-11-27 bc9b019
set quiet = FALSE by default
fix #146
