+++ man/oe_vectortranslate.Rd
@@ -12,7 +12,7 @@ oe_vectortranslate(
   extra_tags = NULL,
   force_vectortranslate = FALSE,
   never_skip_vectortranslate = FALSE,
-  quiet = TRUE
+  quiet = FALSE
 )
 }
 \arguments{
