+++ tests/testthat/test-update.R
@@ -2,7 +2,7 @@ test_that("oe_update(): simplest example works", {
   skip_if_offline()
   fake_dir = tempdir()
   oe_get(
-    "its leeds",
+    "ITS Leeds",
     provider = "test",
     download_directory = fake_dir,
     download_only = TRUE
