### 2020-12-02 8acbdbb
normalize path before gdal_utils
