### 2020-12-02 #395 @potterzot
- **Briefly describe any working relationship you have (had) with the package authors.** I have no working relationship with the authors.

Since not submitting to JOSS, this is ignored:
>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](https://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](https://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

First, I wanted to say thank you for submitting this package. I was excited to review it since it pertains to potential future work I am hoping to do. In general, it's very well written and documented. It would be great to add some guidance on contributing. Other than that, there is really only one issue I came across.

#### Small Issues

- Please add a CONTRIBUTING.md or a section to the README that details how to contribute as per the community guidelines.
- There are four warnings when I run `devtools::test()`, but looking at them I think that is expected behavior since the messages reflect tests of mispecified function arguments.

#### Larger Issues

__oe_match(), oe_get(), and level__

I ran into trouble when running `oe_match`. From the help message it seems like when a location matches to multiple levels, I could select the level I want. `oe_match` doesn't take a `level` parameter, but it also doesn't issue a warning about a level parameter. I ended up thinking it was fine and then also trying to use `level = 2` in `oe_get`, which did issue a warning that was couched in other output, but then returned an error that seemed when I read it unrelated to the inclusion of the level. I suggest either changing the wording to make it clear that you cannot select a level or adding a level parameter.

On a related issue, it's hard to know what "highest level" in the message means without reading. It might be clearer to say something like "smallest administrative unit" or "smallest geography", I'm not sure what the best term would be.

Here's the code and output:

```
> yak <- tmaptools::geocode_OSM("Yakima, WA")$coords
> oe_match(yak, provider = "geofabrik", level = 2)

although coordinates are longitude/latitude, st_intersects assumes that they are planar
The input place was matched with multiple geographical areas. Selecting the areas with the highest "level". See the help page associated to the chosen provider for an explanation of the meaning of the "level" field
The input place was matched with multiple geographical areas with the same "level". Selecting the area whose centroid is closer to the input place
although coordinates are longitude/latitude, st_nearest_points assumes that they are planar
$url
[1] "https://download.geofabrik.de/north-america/us/washington-latest.osm.pbf"

$file_size
[1] 162881219

In st_centroid.sfc(sf::st_geometry(matched_zones)) :
  st_centroid does not give correct centroids for longitude/latitude data
```

This selects the 'Washington State' geometry, but it would be nice if it selected the 'United States' geometry since I specified `level = 2`. Or return an error that says that level is not a valid parameter. The `oe_get` function provides a warning message along those lines, but it gets lost in the rest of the output and ultimately throws an error that is hard to relate back to the inclusion of the `level` argument:

```
> wa <- oe_get(yak, level = 2)

although coordinates are longitude/latitude, st_intersects assumes that they are planar
The input place was matched with multiple geographical areas. Selecting the areas with the highest "level". See the help page associated to the chosen provider for an explanation of the meaning of the "level" field
The input place was matched with multiple geographical areas with the same "level". Selecting the area whose centroid is closer to the input place
although coordinates are longitude/latitude, st_nearest_points assumes that they are planar
Warning: The following arguments are probably misspelled: level
The chosen file was already detected in the download directory. Skip downloading.
The corresponding gpkg file was already detected. Skip vectortranslate operations.
Reading layer `lines' from data source `/tmp/RtmpumWNZh/geofabrik_washington-latest.gpkg' using driver `GPKG'
Error in st_sf(x, ..., agr = agr, sf_column_name = sf_column_name) :
  no simple features geometry column present
In st_centroid.sfc(sf::st_geometry(matched_zones)) :

 Error in st_sf(x, ..., agr = agr, sf_column_name = sf_column_name) :
  no simple features geometry column present
 ```

If I don't include the `level = 2` argument in the `oe_get` function call it works perfectly fine:

```
> wa <- oe_get(yak)

although coordinates are longitude/latitude, st_intersects assumes that they are planar
The input place was matched with multiple geographical areas. Selecting the areas with the highest "level". See the help page associated to the chosen provider for an explanation of the meaning of the "level" field
The input place was matched with multiple geographical areas with the same "level". Selecting the area whose centroid is closer to the input place
although coordinates are longitude/latitude, st_nearest_points assumes that they are planar
The chosen file was already detected in the download directory. Skip downloading.
The corresponding gpkg file was already detected. Skip vectortranslate operations.
Reading layer `lines' from data source `/tmp/RtmpumWNZh/geofabrik_washington-latest.gpkg' using driver `GPKG'
Simple feature collection with 1141221 features and 9 fields
geometry type:  LINESTRING
dimension:      XY
bbox:           xmin: -131.6931275 ymin: 41.9738173 xmax: -115.069417 ymax: 55.3538395
geographic CRS: WGS 84
In st_centroid.sfc(sf::st_geometry(matched_zones)) :
  st_centroid does not give correct centroids for longitude/latitude data
```
