### 2020-12-16 #395 @salvafern
Not submitting to JOSS

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](https://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](https://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

Hello @agila5. I took some time to have a look at this, sorry for that. Let me say I enjoyed reviewing your package and that I could learn a couple things from it. I work daily with geospatial data and I hadn't seen before such a nice way to download, transform and cache data with R. I just got one suggestion and a few minor issues.

#### Suggestion

On my first look at the README I thought that `oe_get()` could retrieve geometries from anywhere, and got a bit disappointed when I saw it was not the case. I think the `oe_match()` function deserves some attention in the README as it is really useful to find what geometries are available.

For example, the very first thing I tried was to look for a city I know, as I think most users will do:

oe_get("Ghent")
# Error: String distance between best match and the input place is 2, while the maximum threshold distance is equal to 1. You should increase the max_string_dist parameter, look for a closer match in the chosen provider database or consider using a different match_by variable.
```

But it wasn't until I read the `osmextract` vignette that I found it.

oe_match("Ghent", provider = "bbbike")
# The input place was matched with: Gent
# $url
# [1] "https://download.bbbike.org/osm/bbbike/Gent/Gent.osm.pbf"
#
# $file_size
# [1] 13099190
```

In this case the issue was the provider, but I found it thanks to the `oe_match()` function.

#### Minor issues

* **Contributing guidelines**: not included in the README or in a separated CONTRIBUTING file. Maybe add a reference to the `providers` vignette in the README.

* **Testing**: 4 warnings (See my environment info below)

```r
test-match.R:38: warning: oe_match: sfc_POINT objects
st_centroid does not give correct centroids for longitude/latitude data

test-match.R:43: warning: oe_match: sfc_POINT objects
st_centroid does not give correct centroids for longitude/latitude data

test-match.R:43: warning: oe_match: sfc_POINT objects
st_centroid does not give correct centroids for longitude/latitude data

test-read.R:46: warning: oe_read fails with misspelled arguments
The following arguments are probably misspelled: stringasfactor
```

* **Code style**: I ran `styler:::style_active_pkg()` and it changed 23 out of 25 files. I didn't compare the changes but maybe you want to have a look.

```R
  styler:::style_active_pkg()
  # Using style transformers `styler::tidyverse_style()`
  # Styling  25  files:
  #   R/data.R                              The R.cache package needs to create a directory that will hold cache files. It is convenient to use  ‘C:\Users\salvadorf\AppData\Local\R\R.cache’ because it follows the standard on your operating system and it remains also after restarting R. Do you wish to create the 'C:\Users\salvadorf\AppData\Local\R\R.cache' directory? If not, a temporary directory (C:\Users\SALVAD~1\AppData\Local\Temp\RtmpwneRPU/.Rcache) that is specific to this R session will be used. [Y/n]:
  #   Y
  # i
  # R/download.R                          i
  # R/find.R                              i
  # R/get.R                               i
  # R/globals.R                           i
  # R/match.R                             i
  # R/providers.R                         i
  # R/read.R                              i
  # R/update.R                            i
  # R/utils.R                             i
  # R/vectortranslate.R                   i
  # R/zzz.R                               i
  # tests/testthat.R                      √
  # tests/testthat/test-download.R        i
  # tests/testthat/test-find.R            √
  # tests/testthat/test-get.R             i
  # tests/testthat/test-match.R           i
  # tests/testthat/test-providers.R       i
  # tests/testthat/test-read.R            i
  # tests/testthat/test-update.R          i
  # tests/testthat/test-vectortranslate.R i
  # data-raw/bbbike_zones.R               i
  # data-raw/geofabrik_zones.R            i
  # data-raw/openstreetmap_fr_zones.R     i
  # data-raw/test_zones.R                 i
  # ----------------------------------------
  #   Status	Count	Legend
  # √ 	2	File unchanged.
  # i 	23	File changed.
  # x 	0	Styling threw an error.
  # ----------------------------------------
  #   Please review the changes carefully!
```

- **Citation**: You could add your preferred citation (See https://devguide.ropensci.org/building.html#readme)
- **Continuous Integration**: Apart of the code coverage, are you planning to add a CI system? (https://devguide.ropensci.org/ci.html#how-to-use-continuous-integration)

#### My environment

sessioninfo::platform_info()
# date     2020-12-16
```
