### 2020-12-16 ITSLeeds/osmextract#154 @Robinlovelace
Search for place location if there's no match in the providers
Reproducible example:

``` r
osmextract::oe_get("leeds")
#> No exact matching found for place = leeds. Best match is Laos.
#> Error: String distance between best match and the input place is 3, while the maximum threshold distance is equal to 1. You should increase the max_string_dist parameter, look for a closer match in the chosen provider database or consider using a different match_by variable.
```

Motivation: useful observation from reviewer https://github.com/ropensci/software-review/issues/395#issuecomment-746784728

> For example, the very first thing I tried was to look for a city I know, as I think most users will do:
