### 2020-12-16 7a8da45
Add function for #154
