### 2020-12-17 #395 @agila5
Hi @salvafern, thank you very much for your review and your useful comments. @annakrystalli I will prepare an answer to both reviews in the next weeks!

Happy holidays to all the people that helped me during the review process ❄️ 🎆 🎉
