### 2020-12-28 c5cebed
A slightly different approach for #155 + fix tests
