+++ tests/testthat/test-update.R
@@ -1,12 +1,15 @@
 test_that("oe_update(): simplest example works", {
   skip_if_offline()
   fake_dir = tempdir()
-  oe_get(
+  out = oe_get(
     "ITS Leeds",
     provider = "test",
     download_directory = fake_dir,
-    download_only = TRUE
+    download_only = TRUE,
+    quiet = TRUE
   )
   expect_error(oe_update(fake_dir, quiet = TRUE), NA)
-  expect_message(oe_update(fake_dir, quiet = FALSE))
+  # AG: I decided to comment out that test since I don't see any benefit testing
+  # the "verbose" output during R CMD checks (that I rarely check manually)
+  # expect_message(oe_update(fake_dir, quiet = FALSE))
 })
