+++ tests/testthat/test-vectortranslate.R
@@ -8,20 +8,22 @@ its_pbf = oe_download(
 )
 
 test_that("oe_vectortranslate: simplest examples work", {
-  its_gpkg = oe_vectortranslate(its_pbf)
+  its_gpkg = oe_vectortranslate(its_pbf, quiet = TRUE)
   expect_equal(tools::file_ext(its_gpkg), "gpkg")
 })
 
 test_that("oe_vectortranslate returns file_path is .gpkg exists", {
-  its_gpkg = oe_vectortranslate(its_pbf)
-  new_its_gpkg = oe_vectortranslate(its_pbf)
+  its_gpkg = oe_vectortranslate(its_pbf, quiet = TRUE)
+  new_its_gpkg = oe_vectortranslate(its_pbf, quiet = TRUE)
   expect_equal(its_gpkg, new_its_gpkg)
 })
 
 test_that("oe_vectortranslate adds new tags", {
   its_gpkg = oe_vectortranslate(
     its_pbf,
-    extra_tags = "oneway", force_vectortranslate = TRUE
+    extra_tags = "oneway",
+    force_vectortranslate = TRUE,
+    quiet = TRUE
   )
   expect_match(
     paste(names(sf::st_read(its_gpkg, quiet = TRUE)), collapse = "-"),
@@ -30,7 +32,7 @@ test_that("oe_vectortranslate adds new tags", {
 })
 
 test_that("oe_vectortranslate adds new tags to existing file", {
-  new_its_gpkg = oe_vectortranslate(its_pbf, extra_tags = c("oneway"))
+  new_its_gpkg = oe_vectortranslate(its_pbf, extra_tags = c("oneway"), quiet = TRUE)
   expect_match(
     paste(names(sf::st_read(new_its_gpkg, quiet = TRUE)), collapse = "-"),
     "oneway"
@@ -39,7 +41,12 @@ test_that("oe_vectortranslate adds new tags to existing file", {
 
 test_that("oe_get_keys: simplest examples work", {
   skip_if_offline()
-  itsleeds_gpkg = oe_get("itsleeds", provider = "test", download_only = TRUE)
+  itsleeds_gpkg = oe_get(
+    "itsleeds",
+    provider = "test",
+    download_only = TRUE,
+    quiet = TRUE
+  )
   expect_type(oe_get_keys(itsleeds_gpkg), "character")
 })
 
@@ -50,7 +57,8 @@ test_that("oe_get_keys: returns error with wrong inputs", {
     "itsleeds",
     provider = "test",
     download_only = TRUE,
-    skip_vectortranslate = TRUE
+    skip_vectortranslate = TRUE,
+    quiet = TRUE
   )
   expect_error(oe_get_keys(itsleeds_pbf)) # wrong format
 })
