+++ tests/testthat/test-find.R
@@ -2,7 +2,8 @@ test_that("oe_find: simplest example works", {
   its_leeds_find <- oe_find(
     "ITS Leeds",
     provider = "test",
-    download_if_missing = TRUE
+    download_if_missing = TRUE,
+    quiet = TRUE
   )
   expect_type(its_leeds_find, "character")
 
@@ -10,7 +11,8 @@ test_that("oe_find: simplest example works", {
   its_leeds_find <- oe_find(
     "ITS Leeds",
     provider = "test",
-    download_if_missing = TRUE
+    download_if_missing = TRUE,
+    quiet = TRUE
   )
   expect_type(its_leeds_find, "character")
   expect_length(its_leeds_find, 2)
