+++ tests/testthat/test-get.R
@@ -1,16 +1,20 @@
 test_that("oe_get: simplest examples work", {
-  expect_s3_class(oe_get("ITS Leeds", provider = "test", quiet = TRUE), "sf")
+  expect_s3_class(
+    oe_get("ITS Leeds", provider = "test", quiet = TRUE, download_directory = tempdir()),
+    "sf"
+  )
+  # clean tempdir
+  file.remove(oe_find("ITS Leeds", provider = "test", download_directory = tempdir()))
 })
 
 test_that("vectortranslate is not skipped if force_download is TRUE", {
   # See https://github.com/ITSLeeds/osmextract/issues/144
   # I need to download the following files in a new directory since they could
   # be mixed with previously downloaded files (and hence ruin the tests)
-  my_tempdir = tempdir()
 
   small_its_leeds <- oe_get(
     "ITS Leeds",
-    download_directory = my_tempdir,
+    download_directory = tempdir(),
     vectortranslate_options = c(
     "-f", "GPKG",
     "-overwrite",
@@ -24,12 +28,15 @@ test_that("vectortranslate is not skipped if force_download is TRUE", {
   # Download it again
   its_leeds <- oe_get(
     "ITS Leeds",
-    download_directory = my_tempdir,
+    download_directory = tempdir(),
     force_vectortranslate = TRUE,
     quiet = TRUE
   )
 
   expect_gte(nrow(its_leeds), nrow(small_its_leeds))
+
+  # clean tempdir
+  file.remove(oe_find("ITS Leeds", provider = "test", download_directory = tempdir()))
 })
 
 test_that("can specify path using ~", {
