+++ tests/testthat/test-vectortranslate.R
@@ -4,18 +4,23 @@ its_pbf = oe_download(
   file_url = its_match$url,
   file_size = its_match$file_size,
   download_directory = tempdir(),
-  provider = "test"
+  provider = "test",
+  quiet = TRUE
 )
 
 test_that("oe_vectortranslate: simplest examples work", {
   its_gpkg = oe_vectortranslate(its_pbf, quiet = TRUE)
   expect_equal(tools::file_ext(its_gpkg), "gpkg")
+  file.remove(its_gpkg)
 })
 
 test_that("oe_vectortranslate returns file_path is .gpkg exists", {
   its_gpkg = oe_vectortranslate(its_pbf, quiet = TRUE)
-  new_its_gpkg = oe_vectortranslate(its_pbf, quiet = TRUE)
-  expect_equal(its_gpkg, new_its_gpkg)
+  expect_message(
+    oe_vectortranslate(its_pbf),
+    "Skip vectortranslate operations."
+  )
+  file.remove(its_gpkg)
 })
 
 test_that("oe_vectortranslate adds new tags", {
@@ -29,36 +34,27 @@ test_that("oe_vectortranslate adds new tags", {
     paste(names(sf::st_read(its_gpkg, quiet = TRUE)), collapse = "-"),
     "oneway"
   )
+  file.remove(its_gpkg)
 })
 
 test_that("oe_vectortranslate adds new tags to existing file", {
+  its_gpkg = oe_vectortranslate(its_pbf, quiet = TRUE)
   new_its_gpkg = oe_vectortranslate(its_pbf, extra_tags = c("oneway"), quiet = TRUE)
   expect_match(
     paste(names(sf::st_read(new_its_gpkg, quiet = TRUE)), collapse = "-"),
     "oneway"
   )
+  file.remove(new_its_gpkg)
 })
 
-test_that("oe_get_keys: simplest examples work", {
-  skip_if_offline()
-  itsleeds_gpkg = oe_get(
-    "itsleeds",
-    provider = "test",
-    download_only = TRUE,
-    quiet = TRUE
-  )
+test_that("oe_get_keys: simplest example works", {
+  itsleeds_gpkg = oe_vectortranslate(its_pbf, quiet = TRUE)
   expect_type(oe_get_keys(itsleeds_gpkg), "character")
+
+  file.remove(itsleeds_gpkg)
 })
 
 test_that("oe_get_keys: returns error with wrong inputs", {
   expect_error(oe_get_keys("xxx.gpkg")) # file does not exist
-  skip_if_offline()
-  itsleeds_pbf = oe_get(
-    "itsleeds",
-    provider = "test",
-    download_only = TRUE,
-    skip_vectortranslate = TRUE,
-    quiet = TRUE
-  )
-  expect_error(oe_get_keys(itsleeds_pbf)) # wrong format
+  expect_error(oe_get_keys(its_pbf)) # wrong format
 })
