### 2020-12-28 72bddbc
new tests for the interaction between oe_match and oe_search
