### 2020-12-28 07d6544
Change startup message
Fix #156
