+++ R/get.R
@@ -5,7 +5,7 @@
 #' Check the introductory vignette, the examples and the help pages of the
 #' wrapped functions to understand the details behind all parameters.
 #'
-#' To learn how to use the `query` argument, for example, see the
+#' To learn how to use the `query` argument see, for example, the
 #' [query section of the osmextract vignette](https://itsleeds.github.io/osmextract/articles/osmextract.html#query).
 #'
 #' @param place Description of the geographical area that should be matched with
@@ -46,23 +46,24 @@
 #'   `.osm.pbf` files.
 #' @param max_file_size The maximum file size to download without asking in
 #'   interactive mode. Default: `5e+8`, half a gigabyte.
-#' @param vectortranslate_options Options to pass to the [`sf::gdal_utils()`]
-#'   argument `options`. Set by default. Check Details in the introductory
-#'   vignette and in [oe_vectortranslate()].
-#' @param osmconf_ini The configuration file specifying which columns should be
-#'   in the resulting data frame. See documentation at
-#'   [gdal.org](https://gdal.org/drivers/vector/osm.html). Check Details in
-#'   [oe_vectortranslate()].
+#' @param vectortranslate_options Options passed to the [`sf::gdal_utils()`]
+#'   argument `options`. Set by default. Check details in the introductory
+#'   vignette and the help page of [oe_vectortranslate()].
+#' @param osmconf_ini The configuration file. See documentation at
+#'   [gdal.org](https://gdal.org/drivers/vector/osm.html). Check details in the
+#'   introductory vignette and the help page of [oe_vectortranslate()]. Set by
+#'   default.
 #' @param extra_tags Which additional columns, corresponding to OSM tags, should
-#'   be in the resulting dataset? `FALSE` by default. Check Details at
-#'   [oe_vectortranslate()] and [oe_get_keys()]. Ignored when `osmconf_ini` is
-#'   not `NULL`.
+#'   be in the resulting dataset? `NULL` by default. Check the introductory
+#'   vignette and the help pages of [oe_vectortranslate()] and [oe_get_keys()].
+#'   Ignored when `osmconf_ini` is not `NULL`.
 #' @param force_vectortranslate Boolean. Force the original `.pbf` file to be
 #'   translated into a `.gpkg` file, even if a `.gpkg` with the same name
-#'   already exists? Check Details at [oe_vectortranslate()] .
-#' @param skip_vectortranslate Boolean. If `TRUE` then the function skips all
-#'   vectortranslate operations and it reads (or simply returns the path) of
-#'   the `.osm.pbf` file. `FALSE` by default.
+#'   already exists? Check the introductory vignette and the help page of
+#'   [oe_vectortranslate()].
+#' @param skip_vectortranslate Boolean. If `TRUE`, then the function skips all
+#'   vectortranslate operations and it reads (or simply returns the path) of the
+#'   `.osm.pbf` file. `FALSE` by default.
 #' @param never_skip_vectortranslate Boolean. This is used in case the user
 #'   passed its own `.ini` file or vectortranslate options (since, in those case,
 #'   it's too difficult to determine if an existing `.gpkg` file was generated
@@ -83,7 +84,7 @@
 #' @export
 #'
 #' @details The algorithm that we use for importing an OSM extract data into R
-#'   is divided into 4 steps: 1) Matching the input `place` with the url of a
+#'   is divided into 4 steps: 1) match the input `place` with the url of a
 #'   `.pbf` file; 2) download the `.pbf` file; 3) convert it into `.gpkg` format
[…]

[…]
