+++ README.Rmd
@@ -169,7 +169,7 @@ If the input place does not match any of the existing names in the supported pro
 For example (not run): 
 
 ```{r, eval = FALSE}
-oe_get("Milan")
+oe_get("Milan") # Warning: It will download more than 400MB of data
 #> No exact match found for place = Milan and provider = geofabrik. Best match is Iran.
 #> Checking the other providers.
 #> No exact match found in any OSM provider data. Searching for the location online.
