### 2020-12-29 dc64469
oe_search should be used only when match_by == name
