+++ vignettes/osmextract.Rmd
@@ -13,6 +13,7 @@ knitr::opts_chunk$set(
   comment = "#>", 
   fig.align = "center"
 )
+Sys.setenv(OSMEXT_DOWNLOAD_DIRECTORY = "")
 ```
 
 This vignette provides an introduction to using the package, building on the README which covers installation and our motivations for creating it.
@@ -24,6 +25,7 @@ library(osmextract)
 ```
 
 The first thing to say is: **do not ignore this message**!
+The Open Street Map (OSM) extracts linked with this packages are stored by external providers such as [Geofabrik](https://download.geofabrik.de/), [Bbbike](https://download.bbbike.org/osm/), or [OpenStreetMap.fr](http://download.openstreetmap.fr/). 
 There are important legal considerations that you should be aware of before using OSM data, especially if you are working in a for-profit capacity.
 
 # Legal considerations
@@ -49,9 +51,11 @@ The package is composed of the following main functions:
 1. `oe_get()`: Match, download, translate, and import data, all in one step.
 
 For many users who just want to get OSM data quickly `oe_get()` may be sufficient, as covered in the README.
-We will demonstrate each function in turn.
+We will demonstrate each function in turn. 
+As you can see, the name of the most important functions in this package start with `oe_*`.
+<!-- Mention auto-completion stuff -->
 
-## `oe_providers()`:list providers
+## `oe_providers()`: List providers
 
 `oe_providers()` lists the providers that are currently available with the version of `osmextract` you have installed.
 
@@ -62,7 +66,7 @@ oe_providers()
 Each element in the column `database_name` is a data object that is packaged with `osmextract`. 
 You can read a detailed description of each provider data running, for example, `?geofabrik_zones` or `?bbbike_zones`. 
 
-Perhaps the best known bulk OSM data provider is Geofabrik, represented as `data.frame` in the packaged object `geofabrik_zones`.
+Perhaps the best known bulk OSM data provider is probably Geofabrik, represented as a `data.frame` in the packaged object `geofabrik_zones`.
 
 ```{r}
 class(geofabrik_zones)
@@ -76,7 +80,7 @@ library(sf)
 ```
 
 That gives you access to many functions for working with geographic vector data of the type provided by `osmextract`.
-Each row of data in an `sf` object contains a geometry, representing the area covered by each provider zone, meaning you can plot the data as follows:
+Each row of data in an `sf` object contains a geometry, representing the area covered by a provider zone, meaning you can plot the data as follows:
 
 ```{r}
 par(mar = rep(0.1, 4))
@@ -100,45 +104,39 @@ We are working on adding support for manually selected regions from the BBBike w
 
 Check the ["Comparing the supported OSM providers"](https://itsleeds.github.io/osmextract/articles/providers_comparisons.html) vignette for some simple guidance on how to choose the best provider. 
 
-## `oe_match()`: finding OSM extracts
+## `oe_match()`: Match OSM extracts
 
 `oe_match()` takes in input a string (through the parameter `place`), and it returns a named list of length two with the URL and the size (in bytes) of a `.osm.pbf` file representing a geographical zone stored by one of the supported providers. 
 For example: 
 
 ```{r}
-oe_match("Italy")
+oe_match("Italy", quiet = TRUE)
 oe_match("Leeds", provider = "bbbike", quiet = TRUE)
 ```
 
 The operations for matching a string with a geographical zone are based on the minimum Approximate String Distance (`?adist()`) between the input `place` and one of the fields in the provider's database. 
-By default, the function uses the `name` field and Geofabrik provider. 
[…]
