### 2021-01-05 #395 @agila5
Hi everyone and happy new year 🎉

I'm sorry that I didn't write any update. During the Christmas holidays, I focused on finishing the last parts of the dissertation (that must be submitted in a couple of days). We sketched a partial answer to the second review [here](https://github.com/ITSLeeds/osmextract/pull/155), but I will prepare a more precise and detailed report for both reviews sometimes during the next week. Sorry for the delay.
