### 2021-01-08 ITSLeeds/osmextract#160 @agila5
Add new "level" parameter in oe_match and improve docs
See https://github.com/ITSLeeds/osmextract/issues/35#issuecomment-738000112 and https://github.com/ropensci/software-review/issues/395#issuecomment-737551925
