### 2021-01-11 ITSLeeds/osmextract#154 @agila5
![Rplot2](https://user-images.githubusercontent.com/22221146/104241581-32bf7800-545e-11eb-896b-3f535dd1af5e.png)
