### 2021-01-11 ITSLeeds/osmextract#154 @agila5
![Rplot3](https://user-images.githubusercontent.com/22221146/104243054-4966ce80-5460-11eb-951b-ca1ce9d09f33.png)
