+++ vignettes/osmextract.Rmd
@@ -7,7 +7,7 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-```{r, include = FALSE}
+```{r setup, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>", 
@@ -20,12 +20,12 @@ This vignette provides an introduction to using the package, building on the REA
 
 Loading the package generates important messages about the license associated with OSM data.
 
-```{r setup}
+```{r}
 library(osmextract)
 ```
 
 The first thing to say is: **do not ignore this message**!
-The Open Street Map (OSM) extracts linked with this packages are stored by external providers such as [Geofabrik](https://download.geofabrik.de/), [Bbbike](https://download.bbbike.org/osm/), or [OpenStreetMap.fr](http://download.openstreetmap.fr/). 
+The Open Street Map (OSM) extracts are stored by external providers such as [Geofabrik](https://download.geofabrik.de/), [Bbbike](https://download.bbbike.org/osm/), or [OpenStreetMap.fr](http://download.openstreetmap.fr/). 
 There are important legal considerations that you should be aware of before using OSM data, especially if you are working in a for-profit capacity.
 
 # Legal considerations
@@ -51,9 +51,8 @@ The package is composed of the following main functions:
 1. `oe_get()`: Match, download, translate, and import data, all in one step.
 
 For many users who just want to get OSM data quickly `oe_get()` may be sufficient, as covered in the README.
-We will demonstrate each function in turn. 
-As you can see, the name of the most important functions in this package start with `oe_*`.
-<!-- Mention auto-completion stuff -->
+We will demonstrate each function in turn, following the same order in which they are typically used. 
+As you can see, the name of the most important functions in this package start with `oe_*` prefix, which means that you can easily use auto-completion features (with Rstudio or similar IDE). 
 
 ## `oe_providers()`: List providers
 
@@ -66,7 +65,7 @@ oe_providers()
 Each element in the column `database_name` is a data object that is packaged with `osmextract`. 
 You can read a detailed description of each provider data running, for example, `?geofabrik_zones` or `?bbbike_zones`. 
 
-Perhaps the best known bulk OSM data provider is probably Geofabrik, represented as a `data.frame` in the packaged object `geofabrik_zones`.
+Perhaps, the best known bulk OSM data provider is Geofabrik, represented as a `data.frame` in the packaged object `geofabrik_zones`.
 
 ```{r}
 class(geofabrik_zones)
@@ -99,26 +98,26 @@ plot(st_geometry(bbbike_zones), xlim = c(-2, 10), ylim = c(35, 60), col = "darkr
 
 <img src="https://user-images.githubusercontent.com/22221146/94461461-772e4d00-01ba-11eb-950c-804ad177729f.png" width="100%" />
 
-As shown in the above visualisation of [BBBike.org](https://download.bbbike.org/osm/) zones in Europe, the provider offers rectangular extracts of major cities.
+As shown in the visualisation above of [BBBike.org](https://download.bbbike.org/osm/) zones in Europe, the provider offers rectangular extracts of the major cities.
 We are working on adding support for manually selected regions from the BBBike website (see https://github.com/ITSLeeds/osmextract/issues/100).
 
-Check the ["Comparing the supported OSM providers"](https://itsleeds.github.io/osmextract/articles/providers_comparisons.html) vignette for some simple guidance on how to choose the best provider. 
+Check the ["Comparing the supported OSM providers"](https://itsleeds.github.io/osmextract/articles/providers_comparisons.html) vignette for some simple guidelines on how to choose the best provider. 
 
 ## `oe_match()`: Match OSM extracts
 
[…]
