### 2021-01-14 61dc08e
test ... against the formals in st_as_sf
fix #152
