### 2021-01-14 8cc469e
normalizepath in download_directory
fix #161
