### 2021-01-14 7e4ac59
Remove leading/trailing space in URL
fix #163
