### 2021-01-14 cc8ff55
Test if there is misalignment between query and layer
fix #122
