+++ DESCRIPTION
@@ -10,8 +10,8 @@ Authors@R: c(
     person("Barry", "Rowlingson", role=c("ctb"),
       comment = c(ORCID = "0000-0002-8586-6625"))
     )
-Description: This package is used to download, convert and read Open Street
-    Map data extracts downloaded from several providers. 
+Description: Match, download, convert and read Open Street Map data extracts 
+    obtained from several providers. 
 License: GPL-3
 Encoding: UTF-8
 LazyData: true
