### 2021-01-14 a5a26e3
add contributing guidelines in the README.Rmd
