+++ README.Rmd
@@ -231,5 +231,11 @@ See the [Introducing osmextract vignette](https://itsleeds.github.io/osmextract/
 - [PostGIS](https://www.bostongis.com/PrinterFriendly.aspx?content_name=loading_osm_postgis) is an established spatial database that works well with large OSM datasets
 - Any others? Let us know!
 
+## Contribution
+
+We look very much forward to comments, questions and contributions. 
+If you have any doubt, or if you want to suggest a new approach, or if you want to add a new OSM provider, feel free to create a new issue in the [issue tracker](https://github.com/ITSLeeds/osmextract/issues) or a new [pull request](https://github.com/ITSLeeds/osmextract/pulls). 
+We always try to create the most intuitive user interface and the most informative error messages, but if you think that something is not clear and could have been explained better, please let us know. 
+
 <!-- :) -->
 <!-- :) -->
