### 2021-01-16 55f7901
Add level argument
fix #160
