### 2021-01-17 878aa84
Add sf method to oe_match
Ref #97
