+++ man/oe_match.Rd
@@ -3,6 +3,7 @@
 \name{oe_match}
 \alias{oe_match}
 \alias{oe_match.default}
+\alias{oe_match.sf}
 \alias{oe_match.sfc_POINT}
 \alias{oe_match.numeric}
 \alias{oe_match.character}
@@ -11,6 +12,8 @@
 oe_match(place, ...)
 
 \method{oe_match}{default}(place, ...)
+
+\method{oe_match}{sf}(place, ...)
 
 \method{oe_match}{sfc_POINT}(place, provider = "geofabrik", level = NULL, quiet = FALSE, ...)
 
