+++ NAMESPACE
@@ -6,6 +6,7 @@ S3method(oe_get_keys,sf)
 S3method(oe_match,character)
 S3method(oe_match,default)
 S3method(oe_match,numeric)
+S3method(oe_match,sf)
 S3method(oe_match,sfc_POINT)
 export(oe_download)
 export(oe_download_directory)
