### 2021-01-17 190960b
add sfc_LINESTRING and sfc_POLYGON to oe_match
