+++ man/oe_match.Rd
@@ -4,6 +4,8 @@
 \alias{oe_match}
 \alias{oe_match.default}
 \alias{oe_match.sf}
+\alias{oe_match.sfc_LINESTRING}
+\alias{oe_match.sfc_POLYGON}
 \alias{oe_match.sfc_POINT}
 \alias{oe_match.numeric}
 \alias{oe_match.character}
@@ -14,6 +16,10 @@ oe_match(place, ...)
 \method{oe_match}{default}(place, ...)
 
 \method{oe_match}{sf}(place, ...)
+
+\method{oe_match}{sfc_LINESTRING}(place, ...)
+
+\method{oe_match}{sfc_POLYGON}(place, ...)
 
 \method{oe_match}{sfc_POINT}(place, provider = "geofabrik", level = NULL, quiet = FALSE, ...)
 
