+++ NAMESPACE
@@ -7,7 +7,9 @@ S3method(oe_match,character)
 S3method(oe_match,default)
 S3method(oe_match,numeric)
 S3method(oe_match,sf)
+S3method(oe_match,sfc_LINESTRING)
 S3method(oe_match,sfc_POINT)
+S3method(oe_match,sfc_POLYGON)
 export(oe_download)
 export(oe_download_directory)
 export(oe_find)
