### 2021-01-17 0e8db75
Match any sf/sfc object for #97
