+++ man/oe_match.Rd
@@ -3,7 +3,7 @@
 \name{oe_match}
 \alias{oe_match}
 \alias{oe_match.default}
-\alias{oe_match.sfc_POINT}
+\alias{oe_match.sf}
 \alias{oe_match.numeric}
 \alias{oe_match.character}
 \title{Match input place with a URL}
@@ -12,7 +12,7 @@ oe_match(place, ...)
 
 \method{oe_match}{default}(place, ...)
 
-\method{oe_match}{sfc_POINT}(place, provider = "geofabrik", quiet = FALSE, ...)
+\method{oe_match}{sf}(place, provider = "geofabrik", quiet = FALSE, ...)
 
 \method{oe_match}{numeric}(place, provider = "geofabrik", quiet = FALSE, ...)
 
