### 2021-01-17 fb88c70
Allow multifeature sf objects as place
