### 2021-01-18 a0dc32f
adopt st_combine approach
