+++ man/oe_match.Rd
@@ -4,9 +4,7 @@
 \alias{oe_match}
 \alias{oe_match.default}
 \alias{oe_match.sf}
-\alias{oe_match.sfc_LINESTRING}
-\alias{oe_match.sfc_POLYGON}
-\alias{oe_match.sfc_POINT}
+\alias{oe_match.sfc}
 \alias{oe_match.numeric}
 \alias{oe_match.character}
 \title{Match input place with a URL}
@@ -17,11 +15,7 @@ oe_match(place, ...)
 
 \method{oe_match}{sf}(place, ...)
 
-\method{oe_match}{sfc_LINESTRING}(place, ...)
-
-\method{oe_match}{sfc_POLYGON}(place, ...)
-
-\method{oe_match}{sfc_POINT}(place, provider = "geofabrik", level = NULL, quiet = FALSE, ...)
+\method{oe_match}{sfc}(place, provider = "geofabrik", level = NULL, quiet = FALSE, ...)
 
 \method{oe_match}{numeric}(place, provider = "geofabrik", quiet = FALSE, ...)
 
