### 2021-01-18 8fe727a
suppress some messages and warnings in oe_match.sfc
cc @Robinlovelace
