+++ R/data.R
@@ -8,7 +8,7 @@
 #' @format An sf object with `r nrow(geofabrik_zones)` rows and
 #' `r ncol(geofabrik_zones)` columns:
 #' \describe{
-#'   \item{id}{A unique identifier, contains letters, numbers and potentially
+#'   \item{id}{A unique identifier. It contains letters, numbers and potentially
 #'   the characters "-" and "/".}
 #'   \item{name}{The, usually English, long-form name of the area.}
 #'   \item{parent}{The identifier of the next larger excerpts that contains this
@@ -54,7 +54,7 @@
 #'   file).}
 #'   \item{geometry}{The sfc for that geographical region. These are not the
 #'   country boundaries but a buffer around countries.}
-#'   \item{pbf_size_size}{Size of the `.pbf` file in bytes.}
+#'   \item{pbf_file_size}{Size of the `.pbf` file in bytes.}
 #' }
 #'
 #' @family provider's-database
@@ -73,7 +73,7 @@
 #'   \item{name}{The, usually English, long-form name of the city.}
 #'   \item{last_modified}{When was it last modified?}
 #'   \item{type}{empty}
-#'   \item{pbf_size_size}{Size of the pbf file in bytes.}
+#'   \item{pbf_file_size}{Size of the pbf file in bytes.}
 #'   \item{base_url}{The base URL for the city.}
 #'   \item{poly_url}{The `.poly` file location.}
 #'   \item{pbf}{Link to the latest `.osm.pbf` file for this region.}
@@ -82,7 +82,7 @@
 #'   matching operations in case of spatial input. The oe_* functions will
 #'   select the geographical area closest to the input place with the highest
 #'   "level". See [geofabrik_zones] for an example of a hierarchical structure.}
-#'   \item{geometry}{The `sfc` for that geographical region, rectangular.}
+#'   \item{geometry}{The `sfg` for that geographical region, rectangular.}
 #' }
 #'
 #' @family provider's-database
@@ -99,17 +99,15 @@
 #' \describe{
 #'   \item{id}{A unique ID for each area. It is used by `oe_update()`.}
 #'   \item{name}{The, usually English, long-form name of the city.}
-#'   \item{parent}{The ID identifier of the next larger excerpts that contains
+#'   \item{parent}{The identifier of the next larger excerpts that contains
 #'   this one, if present.}
 #'   \item{level}{An integer code between 1 and 4. Check
 #'   <http://download.openstreetmap.fr/polygons/> to see the hierarchical
 #'   structure of the zones. 1L correspond to the biggest areas. This is used
-#'   only for matching operations in case of spatial input. The oe_* functions
-#'   will select the geographical area closest to the input place with the
-#'   highest "level"}
+#'   only for matching operations in case of spatial input.}
 #'   \item{pbf}{Link to the latest `.osm.pbf` file for this region.}
-#'   \item{pbf_size_size}{Size of the pbf file in bytes.}
-#'   \item{geometry}{The `sfc`` for that geographical region, rectangular.}
+#'   \item{pbf_file_size}{Size of the pbf file in bytes.}
+#'   \item{geometry}{The `sfg` for that geographical region, rectangular.}
 #' }
 #'
 #' @family provider's-database
