+++ R/download.R
@@ -1,30 +1,30 @@
 #' Download a file given a url
 #'
-#' This function is used to download a file given a url. It focuses on OSM
+#' This function is used to download a file given a URL. It focuses on OSM
 #' extracts with `.osm.pbf` format stored by one of the providers implemented in
-#' the package. The url is specified through the parameter `file_url`.
+#' the package. The URL is specified through the parameter `file_url`.
 #'
 #' @details This function runs several checks before actually downloading a new
 #'   file to avoid overloading the OSM providers. The first step is the
 #'   definition of the file's path associated to the input `file_url`. The path
 #'   is created by pasting together the `download_directory`, the name of chosen
-#'   provider (which may be inferred from the url) and the `basename()` of the
-#'   url. For example, if `file_url` is equal to
+#'   provider (which may be inferred from the URL) and the `basename()` of the
+#'   URL. For example, if `file_url` is equal to
 #'   `"https://download.geofabrik.de/europe/italy-latest.osm.pbf"`, and
 #'   `download_directory = "/tmp"`, then the path is built as
 #'   `"/tmp/geofabrik_italy-latest.osm.pbf"`. Thereafter, the function checks
-#'   the existence of a file in that path and, if that's the case, it returns
-#'   the path. The parameter `force_download` is used to modify this behaviour.
-#'   If there is no file associated with the new path, then the function
-#'   downloads a new file using [download.file()] with `mode = "wb"`, and,
-#'   again, it returns the path.
+#'   the existence of that file and, if it founds it, then it returns the path.
+#'   The parameter `force_download` is used to modify this behaviour. If there
+#'   is no file associated with the new path, then the function downloads a new
+#'   file using [download.file()] with `mode = "wb"`, and, again, it returns the
+#'   path.
 #'
 #' @inheritParams oe_get
-#' @param file_url A url pointing to a `.osm.pbf` file that should be
+#' @param file_url A URL pointing to a `.osm.pbf` file that should be
 #'   downloaded.
-#' @param provider Which provider stores the file that is specified using its
-#'   url? If `NULL` (the default), it is inferred from the url, but it must be
-#'   specified for non-standard cases. See details and examples.
+#' @param provider Which provider stores the file? If `NULL` (the default), it
+#'   may be inferred from the URL, but it must be specified for non-standard
+#'   cases. See details and examples.
 #' @param file_basename The basename of the file. The default behaviour is to
 #'   auto-generate it from the URL using `basename()`.
 #' @param file_size How big is the file? Optional. `NA` by default. If it's
@@ -38,8 +38,7 @@
 #' @examples
 #' its_match = oe_match("ITS Leeds", provider = "test", quiet = TRUE)
 #' # ITS Leeds data are stored on github, which is not a standard provider.
-#' # So we need to specify the provider parameter. See oe_providers() for a
-#' # list of all available providers.
+#' # So we need to specify the provider parameter.
 #' oe_download(
 #'   file_url = its_match$url,
 #'   file_size = its_match$file_size,
@@ -88,6 +87,9 @@ oe_download = function(
     download_directory,
     paste(provider, file_basename, sep = "_")
   )
+
+  # Normalise the file_path
+  file_path = normalizePath(file_path, mustWork = FALSE)
 
   # If the file exists and force_download is FALSE, then raise a message and
   # return the file_path. Otherwise we download it after checking for the
