+++ R/match.R
@@ -1,4 +1,4 @@
-#' Match input place with a URL
+#' Match input place with a url
 #'
 #' This function is used to match an input `place` with the URL of a `.osm.pbf`
 #' file (and its file-size, if present). The URLs are stored in several
@@ -15,26 +15,21 @@
 #'
 #' @seealso [oe_providers()] and [oe_match_pattern()].
 #'
-#' @details The fields `iso3166_1_alpha2` and `iso3166_2` are used by Geofabrik
-#'   provider to perform matching operations using [ISO 3166-1
-#'   alpha-2](https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2) and [ISO
-#'   3166-2](https://en.wikipedia.org/wiki/ISO_3166-2) codes. See
-#'   [geofabrik_zones] for more details.
+#' @details
 #'
-#'   If the input place is specified as a spatial point (either as `sfc_POINT`
-#'   or a pair of numeric coordinates), then the function will return a
-#'   geographical area intersecting the point (or an error if there is no
-#'   intersection between input place and provider's data). The argument `level`
-#'   (which must be specified as an integer between 1 and 4, extreme values
-#'   included) is used to select between multiple geographically nested areas.
-#'   We could roughly say that smaller administrative units correspond to higher
-#'   levels. Check the help page of the chosen provider for more details on
-#'   `level` field. By default, `level = NULL`, which means that `oe_match()`
-#'   will return the area(s) corresponding to the highest level. If there is no
-#'   geographical area at the desired level, then the function will return an
-#'   error. If there are multiple areas at the same `level`` intersecting the
-#'   input place, then the function will return the area whose centroid is
-#'   closer to the input place.
+#'   If the input place is specified as a spatial object (either `sf` or `sfc`),
+#'   then the function will return a geographical area that completely contains
+#'   the object (or an error). The argument `level` (which must be specified as
+#'   an integer between 1 and 4, extreme values included) is used to select
+#'   between multiple geographically nested areas. We could roughly say that
+#'   smaller administrative units correspond to higher levels. Check the help
+#'   page of the chosen provider for more details on `level` field. By default,
+#'   `level = NULL`, which means that `oe_match()` will return the area
+#'   corresponding to the highest available level. If there is no geographical
+#'   area at the desired level, then the function will return an error. If there
+#'   are multiple areas at the same `level` intersecting the input place, then
+#'   the function will return the area whose centroid is closest to the input
+#'   place.
 #'
 #'   If the input place is specified as a character vector and there are
 #'   multiple plausible matches between the input place and the `match_by`
@@ -50,6 +45,12 @@
 #'   will perform a spatial matching operation (see Examples and introductory
 #'   vignette), while, if `match_by != "name"`, then it will return an error.
 #'
+#'   The fields `iso3166_1_alpha2` and `iso3166_2` are used by Geofabrik
+#'   provider to perform matching operations using [ISO 3166-1
+#'   alpha-2](https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2) and [ISO
+#'   3166-2](https://en.wikipedia.org/wiki/ISO_3166-2) codes. See
+#'   [geofabrik_zones] for more details.
+#'
 #' @examples
 #' # The simplest example:
 #' oe_match("Italy")
@@ -57,9 +58,9 @@
 #' # The default provider is "geofabrik", but we can change that:
 #' oe_match("Leeds", provider = "bbbike")
 #'
-#' # By default, the matching operations are performed through the column "name"
[…]
