+++ R/utils.R
@@ -28,7 +28,7 @@ check_layer_provider = function(layer, provider) {
 #'
 #' By default, the download directory is equal to `tempdir()`. You can set a
 #' persistent download directory by adding the following command to your
-#' `.Renviron` file (e.g. with [usethis::edit_r_environ()]):
+#' `.Renviron` file (e.g. with `edit_r_environ` function in `usethis` package:
 #' `OSMEXT_DOWNLOAD_DIRECTORY=/path/to/osm/data`.
 #'
 #' @return A character vector representing the path for the download directory
