### 2021-01-19 #395 @agila5
Dear @potterzot and @salvafern, thank you very much for your reviews. You provided friendly comments and constructive criticisms that, I'm sure, helped us improve our package defining new functions and useful arguments. In the following parts of this response, I will provide several answers to all your comments.

> Please add a CONTRIBUTING.md or a section to the README that details how to contribute as per the community guidelines.

> Contributing guidelines: not included in the README or in a separated CONTRIBUTING file. Maybe add a reference to the providers vignette in the README.

We added the following text at the end of the [README](https://github.com/ITSLeeds/osmextract): _We very much look forward to comments, questions and contributions. If you have any doubt, or if you want to suggest a new approach or add a new OSM provider, feel free to create a new issue in the [issue tracker](https://github.com/ITSLeeds/osmextract/issues) or a new [pull request](https://github.com/ITSLeeds/osmextract/pulls). We always try to build the most intuitive user interface and write the most informative error messages, but if you think that something is not clear and could have been explained better, please let us know._

> There are four warnings when I run devtools::test(), but looking at them I think that is expected behavior since the messages reflect tests of mispecified function arguments.

> Testing: 4 warnings (See my environment info below)

We decided to [suppress some warning messages](https://github.com/ITSLeeds/osmextract/commit/8fe727ab75be11d1af95d5cb88fe746392721a0e) related to spatial operations in `oe_match()`, while the remaining ones are expected since they are related to tests regarding the warning messages.

> I ran into trouble when running oe_match. From the help message, it seems like when a location matches to multiple levels, I could select the level I want. oe_match doesn't take a level parameter, but it also doesn't issue a warning about a level parameter. I ended up thinking it was fine and then also trying to use level = 2 in oe_get, which did issue a warning that was couched in other output, but then returned an error that seemed when I read it unrelated to the inclusion of the level. I suggest either changing the wording to make it clear that you cannot select a level or adding a level parameter.

We recently added a new `level` parameter to choose between multiple geographically nested areas. The default behaviour is the same as before, i.e. it selects the area associated with the highest level, which is linked with the smallest administrative unit:

``` r
yak = c(-120.51084, 46.60156)
osmextract::oe_match(yak)
#> The input place was matched with multiple geographical areas.
#> Selecting the smallest administrative unit. Check ?oe_match for more details.
#> The input place was matched with multiple zones at the same level. Check ?oe_match for more details.
#> Selecting the area whose centroid is closest to the input place.
#> although coordinates are longitude/latitude, st_nearest_points assumes that they are planar
#> $url
#> [1] "https://download.geofabrik.de/north-america/us/washington-latest.osm.pbf"
#>
#> $file_size
#> [1] 180538717
```

The `level` argument can be used to choose a bigger administrative area:

``` r
yak = c(-120.51084, 46.60156)
osmextract::oe_match(yak, level = 1)
#> The input place was matched with multiple geographical areas.
#> Selecting the desired level.
#> $url
#> [1] "https://download.geofabrik.de/north-america-latest.osm.pbf"
#>
#> $file_size
#> [1] 10776557870
```

It returns an error message if there is no area with the chosen level.

``` r
yak = c(-120.51084, 46.60156)
osmextract::oe_match(yak, level = 3)
#> The input place was matched with multiple geographical areas.
#> Selecting the desired level.
#> Error in oe_match.sfc(place, provider = provider, quiet = quiet, ...): The input place does not intersect any area at the chosen level.
```

> On a related issue, it's hard to know what "highest level" in the message means without reading. It might be clearer to say something like "smallest administrative unit" or "smallest geography", I'm not sure what the best term would be.

> The oe_get function provides a warning message along those lines, but it gets lost in the rest of the output and ultimately throws an error that is hard to relate back to the inclusion of the level argument:

We decided to suppress some warnings and simplify some messages. For example, the current output could be something like:

``` r
#> The input place was matched with multiple geographical areas.
#> Selecting the smallest administrative unit. Check ?oe_match for more details.
#> The input place was matched with multiple zones at the same level. Check ?oe_match for more details.
#> Selecting the area whose centroid is closest to the input place.
```
 Do you think it's better now? We did not include many details in the message but just added a reference to the help pages.

> On my first look at the README I thought that oe_get() could retrieve geometries from anywhere, and got a bit disappointed when I saw it was not the case.

We changed the behaviour of `oe_match()` (and, as a consequence, `oe_get()`). Now, if the approximate string distance between the closest string match and the input `place` is greater than `max_string_dist`, then `oe_match()` will also check the other supported providers. For example:

``` r
osmextract::oe_match("Madrid")
#> No exact match found for place = Madrid and provider = geofabrik. Best match is Mali.
#> Checking the other providers.
#> An exact string match was found using provider = bbbike.
#> $url
#> [1] "https://download.bbbike.org/osm/bbbike/Madrid/Madrid.osm.pbf"
#>
#> $file_size
#> [1] 33969933
```

We added a simple interface to Nominatim that enables `oe_get()` to 'geolocate' text strings that cannot be found in the providers. If the input `place` is not perfectly matched with any of the supported providers and `match_by` argument is equal to `"name"`, then `oe_match()` will now, after https://github.com/ITSLeeds/osmextract/pull/155,  use the Nominatim API to geolocate the input `place` and perform a spatial matching operation:

``` r
osmextract::oe_match("Ghent")
#> Warning: The input place was matched with multiple geographical zones: Ghana -
#> Kent. Selecting the first match.
#> No exact match found for place = Ghent and provider = geofabrik. Best match is Ghana.
#> Checking the other providers.
#> No exact match found in any OSM provider data. Searching for the location online.
#> The input place was matched with multiple geographical areas.
#> Selecting the smallest administrative unit. Check ?oe_match for more details.
#> $url
#> [1] "https://download.geofabrik.de/europe/belgium-latest.osm.pbf"
#>
#> $file_size
#> [1] 423345383
```

> Code style: I ran styler:::style_active_pkg() and it changed 23 out of 25 files. I didn't compare the changes but maybe you want to have a look.

I think that this is simply caused by the fact that we use the `=` operator for assignment instead of `<-`.

> Citation: You could add your preferred citation (See https://devguide.ropensci.org/building.html#readme)

We are working on a paper and we will add citation information as soon as possible.

> Continuous Integration: Apart of the code coverage, are you planning to add a CI system?

We use Github Actions: https://github.com/ITSLeeds/osmextract/actions. [Here](https://github.com/ITSLeeds/osmextract/tree/master/.github/workflows) you can browse the corresponding `.yaml` files.
