### 2021-01-26 #395 @salvafern
I overlooked this, sorry! @agila5 Both the new behaviour of `oe_match()` and the Nominatim API work like a charm, thanks! I don't have further comments and I recommend this package.
