### 2021-01-27 #395 @agila5
Hi @annakrystalli and thank you very much for your assistance and for inviting me to join rOpenSci.

@robinlovelace I tried to transfer the ownership of `osmextract` repo but it looks like I don't have enough permissions. I think that you create that repository several months ago so probably you are the only one who can transfer its ownership.

Last but not least, @potterzot and @salvafern if you agree I would like to add you to package's description as reviewers. Thank you very much again for your comments.
