### 2021-01-27 9b704b8
update link in test_zones object
