+++ tests/testthat/test-read.R
@@ -22,7 +22,7 @@ test_that("oe_read: simplest examples work", {
 
 test_that("or_read: simplest example with a URL works", {
   skip_if_offline()
-  my_url = "https://github.com/ITSLeeds/osmextract/raw/master/inst/its-example.osm.pbf"
+  my_url = "https://github.com/ropensci/osmextract/raw/master/inst/its-example.osm.pbf"
   expect_error(
     oe_read(
       my_url,
