+++ R/zzz.R
@@ -3,7 +3,7 @@
     "Data (c) OpenStreetMap contributors, ODbL 1.0. https://www.openstreetmap.org/copyright.",
     "Check the package website, itsleeds.github.io/osmextract, for more details.",
     # todo: update site link on rOpenSci
-    # See https://github.com/ITSLeeds/osmextract/issues/156
+    # See https://github.com/ropensci/osmextract/issues/156
     # "Any product made from OpenStreetMap must cite OSM as the data source.",
     # "Geofabrik data are taken from https://download.geofabrik.de/",
     # "For usage details of bbbike data see https://download.bbbike.org/osm/",
