+++ DESCRIPTION
@@ -3,7 +3,7 @@ Type: Package
 Title: Download and Read OpenStreetMap Data Extracts
 Version: 0.1.0
 Authors@R: c(
-    person("Andrea", "Gilardi", email = "a.gilardi5@campus.unimib.it", role=c("aut", "cre"),
+    person("Andrea", "Gilardi", email = "andrea.gilardi@unimib.it", role=c("aut", "cre"),
       comment = c(ORCID = "0000-0002-9424-7439")),
     person("Robin", "Lovelace", role = c("aut"),
       comment = c(ORCID = "0000-0001-5679-6536")),
