### 2021-01-27 1f15a44
Add templates
@Robinlovelace, what do you think abot this?
