+++ README.Rmd
@@ -22,6 +22,8 @@ knitr::opts_chunk$set(
 [![R build status](https://github.com/ropensci/osmextract/workflows/R-CMD-check/badge.svg)](https://github.com/ropensci/osmextract/actions)
 [![Codecov test coverage](https://codecov.io/gh/ropensci/osmextract/branch/master/graph/badge.svg)](https://codecov.io/gh/ropensci/osmextract?branch=master)
 [![peer-review](https://badges.ropensci.org/395_status.svg)](https://github.com/ropensci/software-review/issues/395)
+[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+
 <!-- badges: end -->
 
 The goal of `osmextract` is to make it easier for people to access OpenStreetMap (OSM) data for reproducible research.
@@ -29,7 +31,7 @@ OSM data is the premier source of freely available, community created geographic
 We aim to enable you to extract it for data-driven work in the public interest.
 
 `osmextract` finds, downloads, converts and imports bulk OSM data hosted by providers such as [Geofabrik GmbH](http://download.geofabrik.de) and [bbbike](https://download.bbbike.org/osm/).
-For information on alternative providers and how to add them see the [providers vignette](https://itsleeds.github.io/osmextract/articles/providers.html).
+For information on alternative providers and how to add them see the [providers vignette](https://docs.ropensci.org/osmextract/articles/providers.html).
 
 ## Why osmextract?
 
@@ -175,7 +177,7 @@ oe_get("Milan") # Warning: It will download more than 400MB of data
 #> ... (extra messages here)
 ```
 
-For further details on using the package, see the [Introducing osmextract vignette](https://itsleeds.github.io/osmextract/articles/osmextract.html).
+For further details on using the package, see the [Introducing osmextract vignette](https://docs.ropensci.org/osmextract/articles/osmextract.html).
 
 ## Persistent download directory
 
@@ -204,7 +206,7 @@ This setting affects all calls to [download.file()](https://www.rdocumentation.o
 ## Next steps
 
 We hope to make the user interface to the SQL syntax more user friendly.
-We would love to see more providers added (see the [Add new OpenStreetMap providers](https://itsleeds.github.io/osmextract/articles/providers.html) for details) and see what people can do with OSM datasets of the type provided by this package in a reproducible and open statistical programming environment for the greater good.
+We would love to see more providers added (see the [Add new OpenStreetMap providers](https://docs.ropensci.org/osmextract/articles/providers.html) for details) and see what people can do with OSM datasets of the type provided by this package in a reproducible and open statistical programming environment for the greater good.
 Any contributions to support this or any other improvements to the package are very welcome via our issue tracker.
 
 ## Licence
@@ -216,7 +218,7 @@ We hope this package will provide easy access to OSM data for reproducible resea
 - i. This License;
 - ii. A later version of this License similar in spirit to this
 
-See the [Introducing osmextract vignette](https://itsleeds.github.io/osmextract/articles/osmextract.html) for more details.
+See the [Introducing osmextract vignette](https://docs.ropensci.org/osmextract/articles/osmextract.html) for more details.
 
 ## Other approaches
 
