### 2021-01-27 #395 @agila5
Dear @annakrystalli, first of all thank you very much again for your support during the review process.

A few hours ago @Robinlovelace transferred the repository to rOpenSci and I think I just completed the steps you mentioned in your previous comments (and I will add @potterzot as reviewer as soon as he agrees). In the next weeks we will also add a short introduction presenting the main functionalities in our package.
