+++ DESCRIPTION
@@ -10,7 +10,9 @@ Authors@R: c(
     person("Barry", "Rowlingson", role=c("ctb"),
       comment = c(ORCID = "0000-0002-8586-6625")), 
     person("Salva", "Fernández", role = c("rev"), 
-    comment = c("Salva reviewed the package (v. 0.1) for rOpenSci, see <https://github.com/ropensci/software-review/issues/395>"))
+    comment = c("Salva reviewed the package (v. 0.1) for rOpenSci, see <https://github.com/ropensci/software-review/issues/395>")), 
+    person("Nicholas", "Potter", role = c("rev"), 
+    comment = c("Nicholas reviewed the package (v. 0.1) for rOpenSci, see <https://github.com/ropensci/software-review/issues/395>"))
     )
 Description: Find, download, convert and read Open Street Map data extracts 
     obtained from several providers. 
