+++ DESCRIPTION
@@ -21,7 +21,7 @@ Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.1.1
 Roxygen: list(markdown = TRUE)
-URL: https://github.com/ropensci/osmextract
+URL: https://docs.ropensci.org/osmextract (website) https://github.com/ropensci/osmextract
 BugReports: https://github.com/ropensci/osmextract/issues
 Depends: R (>= 3.5.0)
 Imports: 
