### 2021-01-30 #395 @agila5
Dear @stefaniebutland, thank you very much. I'm pleased to write a blog post for rOpenSci introducing `osmextract` and the Open Street Map ecosystem. As I said to Mark, I hope to write the first draft during the second half of January.

> When you're ready, please submit a draft post by pull request, following instructions in https://blogguide.ropensci.org/. Then my colleague @steffilazerte will review it and suggest a publication date. Cheers!

Ok!
