### 2021-02-01 #395 @annakrystalli
Thanks @agila5. If everything is now complete I will go ahead and close this issue now.

I have also made you full admin of the repo again too. Look forward to the blogpost!

One last thing. I just noticed that the package version is still the same as that submitted (0.1.0). Given the changes that have been made, the version should now be different. You might want to add a [`NEWS.md`](https://devguide.ropensci.org/releasing.html#versioning) file to keep track of these and future changes associated with new package versions.
