+++ NEWS.md
@@ -1,0 +1,5 @@
+osmextract 0.1.0
+=========================
+
+* Finish development of the main functions
+* Submit to rOpenSci for peer-review
