### 2021-02-01 d64fd4f
first draft news for 0.2.0
