+++ NEWS.md
@@ -1,3 +1,27 @@
+osmextract 0.2.0 - in development 
+=========================
+
+### NEW FEATURES
+
+* Add a `level` parameter to `oe_match()`. It is used to choose between multiple hierarchically nested OSM extracts. The default behaviour is to select the smallest administrative unit. (#160)
+* Modify the behaviour of `oe_match()`. The function checks all implemented providers in case the input `place` is not matched with any geographical zone for the chosen provider.
+* Add a simple interface to Nominatim API that enables `oe_match()` to geolocate text strings that cannot be found in the providers. 
+
+### MINOR IMPROVEMENTS
+
+* Normalise the paths managed by `oe_download_directory` and `oe_download` (#150 and #161) 
+
+### BUG FIXES
+
+* Fix the structure of `geofabrik_zones` object (#167)
+* Fix warning messages related to ... in `oe_get()` (#152)
+
+### DOCUMENTATION FIXES
+
+* Simplify several warning messages in case of spatial matching
+* Simplify startup message (#156)
+* Add more details related to download timeouts (#145)
+
 osmextract 0.1.0
 =========================
 
