### 2021-02-02 59b1994
A more informative error message in oe_get_keys
