+++ R/vectortranslate.R
@@ -446,7 +446,11 @@ oe_get_keys.character = function(zone, layer = "lines") {
   }
 
   if (!file.exists(zone)) {
-    stop("The input file does not exist.", call. = FALSE)
+    stop(
+      "The input file does not exist.",
+      "You can download it using oe_get(zone, download_only = TRUE)."
+      call. = FALSE
+    )
   }
 
   if (tools::file_ext(zone) != "gpkg") {
