### 2021-02-02 943aa49
oe_get_keys returns an informative error when there is no other_tags field in the input file
fix #158
