### 2021-02-02 75d903c
simplify templates
