### 2021-02-02 #395 @agila5
Dear @annakrystalli, thank you very much for your suggestions. We just released a [new version](https://github.com/ropensci/osmextract/releases) and created the [NEWS file](https://docs.ropensci.org/osmextract/news/index.html).
