### 2021-02-02 #395 @annakrystalli
Fantastic! Don't forget to change the version on your [`DESCRIPTION`](https://github.com/ropensci/osmextract/blob/master/DESCRIPTION) file 😉
