### 2021-02-02 #395 @Robinlovelace
```
usethis::use_version()
```

Does it automatically ICYMI :tada:
