### 2021-02-02 5908c07
typo in osmextract vignette
